% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{brkAnnual}
\alias{brkAnnual}
\title{Berkshire Hathaway annual return}
\source{
TBA
}
\usage{
data("brkAnnual")
}
\description{
Berkshire Hathaway annual return, with year end price and total return
}
