% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{brkMonthlyBloomberg}
\alias{brkMonthlyBloomberg}
\title{Berkshire Hathaway montly return}
\source{
Bloomberg
}
\usage{
data("brkMonthlyBloomberg")
}
\description{
Berkshire Hathaway montly return
}
