% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFfmDT.R
\name{calcFLAM}
\alias{calcFLAM}
\title{calcFLAM}
\usage{
calcFLAM(
  specObj,
  modelStats,
  fitResults,
  analysis = c("ISM", "NEW"),
  targetedVol = 0.06,
  ...
)
}
\arguments{
\item{specObj}{an object as the output from specFfm function}

\item{modelStats}{output of the extractRegressionStats functions.
Contains fit statistics of the factor model.}

\item{fitResults}{output from fitFfmDT}

\item{analysis}{type character, choice of c("none", "ISM","NEW"). Default = "none".
Corresponds to methods used in the analysis of fundamental law of active management.}

\item{targetedVol}{numeric; the targeted portfolio volatility in the analysis.
Default is 0.06.}

\item{...}{additional arguments}
}
\description{
function to calculate fundamental law of active management
}
