% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartCusum.R
\name{chartCusum}
\alias{chartCusum}
\title{cusumActMgr Plots}
\usage{
chartCusum(object, digits = 3, which = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{cusumActMgr} returned by \code{cusumActMgr}.}

\item{digits}{The number of digits of numerical values in graphs}

\item{which}{A number or a vector of numbers to indicate the type of plots.
If a subset of the plots is required, specify a subset of the numbers 1 and 2
for plots. The numbers 1 and 2 represent: \cr \cr
1 = Barplot of log-excess returns with annually moving average returns,
plot of tracking error, Plot of excess volatility, and Plot of cusum for returns
with protractors of slopes representing the annualized returns,\cr
2 = Barplot of the information ratio, plot of cusumIR with protractors
of slopes with corresponding IR, and a plot of Lindley's Recursion with
the thresholds, and a scatter plot between the fund and benchmark returns
with robust regression\cr}

\item{...}{other graphics parameters in plot}
}
\value{
Graph(s) as specified by the user.
}
\description{
Plot the elements of an \code{cusumActMgr} object. The EWMA tracking errors are computed
month-to-month by the simple weighted average of the "log-excess returns" volatility and its former (adjacent) time period volatility. 
The excess volatility is the EWMA tracking error difference from the portfolio returns EWMA tracking error
and the benchmark returns EWMA tracking error. Annualized IR is computed by using the monthly log-excess returns
and the annualized EWMA tracking errors.
}
\examples{
data("cusumData")
results = cusumActMgr(portfolioName = "Parvest", benchmarkName = "RUS2500", data = cusumData)
chartCusum(results, which = 1)
chartCusum(results, which = c(1,2))
}
\author{
Chindhanai Uthaisaad
}
