% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartRobRisk.R
\name{chartRobRisk}
\alias{chartRobRisk}
\title{\code{riskBudget} Plots}
\usage{
chartRobRisk(object, digits = 3, col = c(4, 2), ...)
}
\arguments{
\item{object}{An object of class \code{riskBudget} returned by \code{robRiskBudget}.}

\item{digits}{The number of digits of numerical values in graphs}

\item{col}{The vector of numerics representing the color of the initial and final risk budgetings, respectively.}

\item{...}{other graphics parameters in plot}
}
\value{
The graph of initial risk budgets and final risk budgets
}
\description{
Plot the elements of an \code{riskBudget} object.
}
\examples{
data("RussellData")
rf = RussellData[, 16]
robRiskData = RussellData[, 1:15]

riskBudget = robRiskBudget(robRiskData, rf = rf, shrink = TRUE, avgCor = TRUE,
ESMethod = "historical", corMatMethod = "mcd")
chartRobRisk(riskBudget)

}
\author{
Chindhanai Uthaisaad
}
