% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coef.sfm}
\alias{coef.sfm}
\title{Extract coefficients from a fitted stochastic factor model}
\usage{
\method{coef}{sfm}(object, ...)
}
\arguments{
\item{object}{a fit object of class \code{sfm} which is returned by
\code{\link{fitSFM}}}

\item{...}{other arguments passed}
}
\value{
\item{coef.mat}{an N x (K+1) matrix of all coefficients}
where, N is the number of assets and K is the number of factors.
}
\description{
Method or helper function for fit object of class \code{sfm}.
}
\author{
Eric Zivot and Sangeetha Srinivasan
}
\seealso{
\code{\link{fitTSFM}}
}

