% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coef.tsfm}
\alias{coef.tsfm}
\title{Extract coefficients from a fitted time series factor model}
\usage{
\method{coef}{tsfm}(object, ...)
}
\arguments{
\item{object}{a fit object of class \code{tsfm} which is returned by
\code{\link{fitTSFM}}}

\item{...}{other arguments passed}
}
\value{
\item{coef.mat}{an N x (K+1) matrix of all coefficients}
where, N is the number of assets and K is the number of factors.
}
\description{
Method or helper function for fit object of class \code{tsfm}.
}
\examples{
\dontrun{
data(managers.df)
fit <- fitTSFM(asset.names=colnames(managers.df[,(1:6)]),
               factor.names=colnames(managers.df[,7:9]),
               market.name="SP500.TR",
               data=data, fit.method="OLS", variable.selection="none",
               add.up.market=TRUE, add.market.sqd=TRUE)
coef(fit)
}
}
\author{
Eric Zivot and Sangeetha Srinivasan
}
\seealso{
\code{\link{fitTSFM}}
}

