% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFfmDT.R
\name{convert.ffmSpec}
\alias{convert.ffmSpec}
\title{Function to convert to current class # mido to change to retroFit}
\usage{
\method{convert}{ffmSpec}(SpecObj, FitObj, RegStatsObj, ...)
}
\arguments{
\item{SpecObj}{an object as the output from specFfm function}

\item{FitObj}{an object as the output from fitFfmDT function}

\item{RegStatsObj}{an object as the output from extractRegressionStats function}

\item{...}{additional arguments}
}
\description{
Function to convert to current class # mido to change to retroFit
}
