% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFfmDT.R
\name{convert}
\alias{convert}
\title{convert}
\usage{
convert(SpecObj, FitObj, RegStatsObj, ...)
}
\arguments{
\item{SpecObj}{an object as the output from specFfm function}

\item{FitObj}{an object as the output from fitFfmDT function}

\item{RegStatsObj}{an object as the output from extractRegressionStats function}

\item{...}{additional arguments}
}
\description{
function to convert the new ffm spec object to ffm object to make it
easier in plotting and reporting
}
