\name{CRSP}
\alias{CRSP}
\docType{data}
\title{
447 assets from 1996-01-01 through 2003-12-31 from CRSP.
}
\description{
447 assets from 1996-01-01 through 2003-12-31 from CRSP.
}
\details{
Continuous data: 
PRICE, RETURN, VOLUME, SHARES.OUT, MARKET.EQUITY,LTDEBT, NET.SALES, COMMON.EQUITY, NET.INCOME, STOCKHOLDERS.EQUITY, LOG.MARKETCAP, LOG.PRICE, BOOK2MARKET
Categorical data: 
GICS, GICS.INDUSTRY, GICS.SECTOR
Identication data:
DATE, PERMNO, COMPANY.NAME, TICKER
}
\source{
beta package "facmod" from Guy Yullen
}
\references{
Guy Yullen
}
\keyword{datasets}
