% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cusumData}
\alias{cusumData}
\title{Parvest and Russell2500}
\source{
TBA
}
\usage{
data("cusumData")
}
\description{
Data in the example is an xts object containing two monthly returns in each column.
             The first column is the fund returns and the second column is the
             benchmark returns. The data is from Jan 2005 to Dec 2003.
}
