% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmmc.R
\name{.fmmc.boot}
\alias{.fmmc.boot}
\title{#' @title  Statistic function for the boot call.}
\usage{
.fmmc.boot(data, indices, args)
}
\arguments{
\item{data}{matrix of (all factors + returns of just 1 asset)}

\item{indices}{row numbers generated by boot}

\item{args}{additinal paramters needed for subsetting the data and calulating
the perfomance/risk measure.}
}
\description{
Calculates the risk or performance measure by using the estimation function in its argument list.
}
\details{
This method works as follows.
1. Get data with factors and returns.
2. Subset T rows from the data.
3. Discard first TR-TR1 of the asset returns by setting them to NA
4. calls .fmmc.proc method over the new data set to get a new joint empirical
   distribution of returns and factors
5. We use the new returns with the estimation function to calculate the
   risk or performance measure.
}
\author{
Rohit Arora
}
