% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmmc.R
\name{.fmmc.worker}
\alias{.fmmc.worker}
\title{Worker function that acts between the fmmc procedure and calling method.}
\usage{
.fmmc.worker(R, factors, ...)
}
\arguments{
\item{R}{single vector of returns}

\item{factors}{matrix of factor returns}

\item{...}{allows passing paramters to FactorAnalytics.}
}
\description{
Worker function that acts between the fmmc procedure and calling method.
}
\details{
This method takes in data as single time series and factors as xts objects
It then calls the actual estimation procedure.
}
\author{
Rohit Arora
}
