% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposuresTseries.R
\name{exposuresTseries}
\alias{exposuresTseries}
\title{Time series plots of Style Exposures}
\usage{
exposuresTseries(
  data,
  tickers = NULL,
  which.exposures = c("SIZE", "P2B", "ENTVAL"),
  plot.returns = TRUE,
  axis.cex = 1,
  plot.type = "b",
  ...
)
}
\arguments{
\item{data}{Dataframe with time series of style exposures and returns}

\item{tickers}{character or vector of characters of Ticker symbols}

\item{which.exposures}{exposures to plot}

\item{plot.returns}{Logical. If 'TRUE', returns are also plotted along with the exposures.}

\item{axis.cex}{numeric value to control the size of axis labels}

\item{plot.type}{character defining the type of plot. Default is 'b'}

\item{...}{other optional arguments}
}
\value{
Time series plots
}
\description{
Plot the time series plot of returns and style exposures of the data.
}
\examples{

data("factorDataSetDjia5Yrs")

#One ticker with returns plot and one exposure
tickers = "BAC"
exposuresTseries(factorDataSetDjia5Yrs,tickers = tickers,which.exposures = "MKTCAP")

#Multiple Tickers without returns and all exposures.  
tickers = c("AA", "BAC", "IBM")
exposuresTseries(factorDataSetDjia5Yrs,tickers = tickers,plot.returns = FALSE)

#Multiple tickers With returns
exposuresTseries(factorDataSetDjia5Yrs,tickers = tickers,plot.returns = TRUE, axis.cex = 0.8,
plot.type="b")

}
\author{
Avinash Acharya
}
