% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFfmDT.R
\name{extractRegressionStats}
\alias{extractRegressionStats}
\title{extractRegressionStats}
\usage{
extractRegressionStats(specObj, fitResults, full.resid.cov = FALSE)
}
\arguments{
\item{specObj}{fitFM object that has been already fit}

\item{fitResults}{output from fitFMDT}

\item{full.resid.cov}{an option to calculate the full residual covariance or not}
}
\value{
a structure of class ffm holding all the information
}
\description{
function to compute or Extract objects to be returned
}
