% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{factorDataSetDjia}
\alias{factorDataSetDjia}
\title{DJIA Factor Data}
\source{
TBA
}
\usage{
data("factorDataSetDjia")
}
\description{
Contains returns for 30 DJIA stocks along with the factor data (MARKETCAP, ENTVALUE, P2B, EV2S, SIZE)
             starting from Jan 2000 to march 2013.
}

