% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{factorDataSPGMI}
\alias{factorDataSPGMI}
\title{Fundamental factor scores from S&P Global Market Intelligence}
\source{
S&P Global Market Intelligence
}
\usage{
data("factorDataSPGMI")
}
\description{
S&P Global Market Intelligence has kindly provided firm fundamentals data
referred to as scores or alpha factors for educational use in the open source FactorAnalytics
R package. The data is contained in the R data frame object factorDataSPGMI
consisting of the following cross-section of scores for approximately 300 stocks from 1990 to
2015: AccrualRatioCF, AnnVol12M, Beta60M, BP, Chg1YEPS, DivP, EBITDAEV, EP, EQ-style,
LogMktCap, PM12M1M, ROE. This data greatly facilitates the educational value to users of the
fundamental factor model in FactorAnalytics. The package developers wish to thank S&P Global
Market Intelligence for contributing this data to the FactorAnalytics package.
}
