% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorsSPGMI.R
\docType{data}
\name{factorsSPGMI}
\alias{factorsSPGMI}
\title{factorsSPGMI}
\format{
A data.frame containing 14 SPGMI monthly factor exposures (alpha
factors) for 300 stocks from 1993 to 2015 = 276 months of observations on 21
variables that include the 14 factor exposures, for each of 300 stocks
\itemize{
 \item \strong{Date:} type `Date`.
 \item \strong{Ticker:} type `chr`. This is the ticker as of December 2015
 \item \strong{TickerHist:} type `chr`. This is the actual ticker each month
 \item \strong{Company:} type `chr`. The name of company 
 \item \strong{CapGroup:} type `chr`. The market capitalization group of the
 company, one of the following: MicroCap, SmallCap, MidCap or LargeCap
 \item \strong{GICS:} type `chr`. 8 digit S&P GICS code
 \item \strong{Sector:} type `chr`. One of 10 sectors specified by the first
 two digits of the GICS code
 \item \strong{AnnVol12M:} type `num`. Stock nnualized 12 month volatility
 \item \strong{Beta60M:} type `num`. Stock beta computed using 60 months
 \item \strong{BP:} type `num`. Stock book to price ratio
 \item \strong{EP:} type `num`. Stock earnings to price ratio
 \item \strong{LogMktCap:} type `num`. Log of stock market cap in $M
 \item \strong{PM12M1M:} type `num`. Stock return from previous 12 months to
 previous 1 month
 \item \strong{CFROIC:} type `num`. Ratio of company's cash flow to return on
 invested capital
 \item \strong{Chg1YAstTo:} type `num`. One year change in sales turnover
 \item \strong{EBITDAEV:} type `num`. Ratio of EBITDA to enterprise value
 \item \strong{FCFP:} type `num`. Ratio of free cash flow to price
 \item \strong{PM1M:} type `num`. Stock return over previous one month
 \item \strong{SEV:} type `num`. Ratio of sales to enterprise value
}
}
\source{
Standard and Poors Global Market Intelligence (SPGMI). NOTE: SPGMI
data is not covered by the GPL. Redistribution of the SPGMI data is not
permitted, and use of the data in derivative works is not permitted without
the written permission of CRSP.
}
\usage{
data('factorsSPGMI')
}
\description{
14 SPGMI monthly alpha factors for 300 stocks from 1993 to 2015
}
\examples{
 
data(factorsSPGMI)
str(factorsSPGMI)
head(factorsSPGMI, 5)
}
\keyword{datasets}
