% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmRsq.R
\name{ffmRsq}
\alias{ffmRsq}
\title{Factor Model R-squared and VIF Values}
\usage{
ffmRsq(ffmObj, rsq = T, rsqAdj = F, VIF = F, plt.type = 2, digits = 2,
  isPrint = T, isPlot = T, lwd = 2, title = TRUE, ...)
}
\arguments{
\item{ffmObj}{an object of class \code{ffm} produced by \code{fitFfm}}

\item{rsq}{logical; if \code{TRUE}, Factor Model R-squared values are computed for the portfolio. Default is \code{TRUE}.}

\item{rsqAdj}{logical; if \code{TRUE}, Adjusted R-squared values are computed for the portfolio. Default is \code{FALSE}.}

\item{VIF}{logical; if \code{TRUE}, Variance Inflation factor is calculated. Default is \code{FALSE}.
At least 2 continous variables are required in \code{exposure.vars} of fitted model to find VIF.}

\item{plt.type}{a number to indicate the type of plot for plotting Factor Model R-squared/Adj. R-squared values.
1 indicates barplot, 2 indicates time series xy plot. Default is 2.}

\item{digits}{an integer indicating the number of decimal places to be used for rounding. Default is 2.}

\item{isPrint}{logical. if \code{TRUE}, the time series of the computed factor model values is printed along with their mean values.
Else, only the mean values are printed. Default is \code{TRUE}.}

\item{isPlot}{logical. if \code{TRUE}, the time series of the output is plotted. Default is \code{TRUE}.}

\item{lwd}{line width relative to the default. Default is 2.}

\item{title}{logical. if \code{TRUE}, the plots will have the main tiltle. default is \code{TRUE}.}

\item{...}{potentially further arguments passed.}
}
\value{
\code{ffmRsq} returns the sample mean values and plots the time series of corresponding R squared values
                        and the Variance Inflation factors depending on the values of \code{rsq}, \code{rsqAdj} and \code{VIF}.
                        The time series of the output values are also printed if \code{isPrint} is \code{TRUE}
}
\description{
Calcluate and plot the Factor Model R-squared, Adjusted R-squared and Variance Inflation Factors for a portfolio of assets
}
\examples{

#Load the data
 data("factorDataSetDjia5Yrs")

#Fit a Ffm
require(factorAnalytics)
 fit <- fitFfm(data=factorDataSetDjia5Yrs, asset.var="TICKER", ret.var="RETURN",
              date.var="DATE", exposure.vars="SECTOR")

#Calcuate and plot the portfolio R-squared values
 ffmRsq(fit)
 
 fit1 <- fitFfm(data=factorDataSetDjia5Yrs, asset.var="TICKER", ret.var="RETURN",
              date.var="DATE", exposure.vars=c("SECTOR", "P2B", "EV2S", "MKTCAP"))

#Plot and print the time series of Adj R-squared and VIF values
 ffmRsq(fit1, VIF=TRUE, rsqAdj=TRUE, isPrint=TRUE, plt.type = 2)
}
\author{
Doug Martin, Avinash Acharya
}

