% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fitSfm.R
\name{fitSfm}
\alias{coef.sfm}
\alias{fitSfm}
\alias{fitted.sfm}
\alias{residuals.sfm}
\title{Fit a statistical factor model using principal component analysis}
\usage{
fitSfm(data, k = 1, max.k = NULL, refine = TRUE, sig = 0.05,
  check = FALSE, corr = FALSE, ...)

\method{coef}{sfm}(object, ...)

\method{fitted}{sfm}(object, ...)

\method{residuals}{sfm}(object, ...)
}
\arguments{
\item{data}{vector, matrix, data.frame, xts, timeSeries or zoo object with
asset returns. See details.}

\item{k}{number of factors; a number (or) a method for determining the
optimal number of factors, one of "bn" or "ck". See details. Default is 1.}

\item{max.k}{scalar; the maximum number of factors to be considered for
methods "bn" or "ck". Default is \code{NULL}. See details.}

\item{refine}{logical; whether to use the Connor-Korajczyk refinement for
APCA. Default is \code{TRUE}.}

\item{sig}{scalar; desired level of significance when "ck" method is
specified. Default is 0.05.}

\item{check}{logical; to check if any asset has identical observations.
Default is \code{FALSE}.}

\item{corr}{logical; whether to use the correlation instead of the covariance
matrix when finding the principal components. Default is \code{FALSE}.}

\item{...}{optional arguments passed to \code{\link[stats]{lm}}.}

\item{object}{a fit object of class \code{sfm} which is returned by
\code{fitSfm}}
}
\value{
fitTsfm returns an object of class \code{"sfm"} for which
\code{print}, \code{plot}, \code{predict} and \code{summary} methods exist.

The generic accessor functions \code{coef}, \code{fitted} and
\code{residuals} extract various useful features of the fit object.
Additionally, \code{fmCov} computes the covariance matrix for asset returns
based on the fitted factor model

An object of class \code{"sfm"} is a list containing the following
components:
\item{asset.fit}{fitted object of class \code{"mlm"} or \code{"lm"} from the
time-series OLS regression of asset returns on estimated factors.}
\item{k}{number of factors; as input or determined by "ck" or "bn" methods.}
\item{factors}{T x K xts object of estimated factor realizations.}
\item{loadings}{N x K matrix of factor loadings estimated by
regressing the asset returns on estimated factors.}
\item{alpha}{length-N vector of estimated alphas.}
\item{r2}{length-N vector of R-squared values.}
\item{resid.sd}{length-N vector of residual standard deviations.}
\item{residuals}{T x N xts object of residuals from the OLS regression.}
\item{Omega}{N x N return covariance matrix estimated by the factor model.}
\item{eigen}{length-N (or length-T for APCA) vector of eigenvalues of the
sample covariance matrix.}
\item{mimic}{N x K matrix of factor mimicking portfolio weights.}
\item{call}{the matched function call.}
\item{data}{T x N xts data object containing the asset returns.}
\item{asset.names}{length-N vector of column names from data.}
Where N is the number of assets, K is the number of factors, and T is the
number of observations.
}
\description{
Fits a statistical factor model using Principal Component
Analysis (PCA) for one or more asset returns or excess returns. When the
number of assets exceeds the number of time periods, Asymptotic Principal
Component Analysis (APCA) is performed. An object of class \code{"sfm"} is
returned. This function is based on the S+FinMetric function \code{mfactor}.
}
\details{
If \code{data} is not of class \code{"xts"}, rownames must provide an
\code{"xts"} compatible time index. Before model fitting, incomplete cases in
\code{data} are removed using \code{\link[stats]{na.omit}}. Specifying
\code{check=TRUE}, issues a warning if any asset is found to have identical
observations.

Let \code{N} be the number of columns or assets and \code{T} be the number
of rows or observations. When \code{N < T}, Principal Component Analysis
(PCA) is performed. Otherwise, Asymptotic Principal Component Analysis
(APCA) is performed. In either case, any number of factors less than
\code{min(N,T)} can be chosen via argument \code{k}. Default is 1. Refer to
Zivot and Wang (2007) for more details and references.

Alternately, for APCA, a method to determine the number of factors can be
specified: \code{k="bn"} corresponds to Bai and Ng (2002) and \code{k="ck"}
corresponds to Connor and Korajczyk (1993). User can specify the maximum
number of factors, \code{max.k} to consider with these methods. If not, it
is assumed to be either 10 or $T-1$, whichever is smaller.

\code{refine} specifies whether a refinement of the APCA procedure from
Connor and Korajczyk (1988) that may improve efficiency is to be used.

When \code{corr=TRUE}, the correlation matrix of returns are used for
finding the principal components instead of the covariance matrix. This is
typically decided by practioners on a case-by-case basis. The variable with
the highest variance dominates the PCA when the covariance matrix is used.
However, this may be justified if a volatile asset is more interesting for
some reason and volatility information shouldn't be discarded. On the other
hand, using the correlation matrix standardizes the variables and makes them
comparable, avoiding penalizing variables with less dispersion.

Finally, if the median of the 1st principal component is negative, all it's
factor realizations are automatically inverted to enable more meaningful
interpretation.
}
\examples{
# load return data
data(StockReturns)

# PCA is performed on r.M and APCA on r.W
class(r.M)
dim(r.M)
range(rownames(r.M))
class(r.W)
dim(r.W)

# PCA
args(fitSfm)
fit.pca <- fitSfm(r.M, k=2)
class(fit.pca)
names(fit.pca)
head(fit.pca$factors)
head(fit.pca$loadings)
fit.pca$r2
fit.pca$resid.sd
fit.pca$mimic

# APCA with number of factors, k=15
fit.apca <- fitSfm(r.W, k=15, refine=TRUE)

# APCA with the Bai & Ng method
fit.apca.bn <- fitSfm(r.W, k="bn")

# APCA with the Connor-Korajczyk method
fit.apca.ck <- fitSfm(r.W, k="ck")
}
\author{
Eric Zivot, Sangeetha Srinivasan and Yi-An Chen
}
\references{
Bai, J., & Ng, S. (2002). Determining the number of factors in approximate
factor models. Econometrica, 70(1), 191-221.

Connor, G., & Korajczyk, R. A. (1988). Risk and return in an equilibrium
APT: Application of a new test methodology. Journal of Financial Economics,
21(2), 255-289.

Connor, G., & Korajczyk, R. A. (1993). A test for the number of factors in
an approximate factor model. The Journal of Finance, 48(4), 1263-1291.

Zivot, E., & Wang, J. (2007). Modeling Financial Time Series with S-PLUS
(Vol. 191). Springer.
}
\seealso{
The \code{sfm} methods for generic functions:
\code{\link{plot.sfm}}, \code{\link{predict.sfm}},
\code{\link{print.sfm}} and \code{\link{summary.sfm}}.

And, the following extractor functions: \code{\link[stats]{coef}},
\code{\link[stats]{fitted}}, \code{\link[stats]{residuals}},
\code{\link{fmCov}}, \code{\link{fmSdDecomp}}, \code{\link{fmVaRDecomp}}
and \code{\link{fmEsDecomp}}.

\code{\link{paFm}} for Performance Attribution.
}

