% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitted.sfm}
\alias{fitted.sfm}
\title{Get fitted values from a stochastic factor model}
\usage{
\method{fitted}{tsfm}(x)
}
\arguments{
\item{x}{an object of class \code{sfm} which is returned by
\code{\link{fitSFM}}}
}
\value{
\item{fitted.xts}{an N x T data object of fitted values}
where, N is the number of assets and T is the number of time periods.
}
\description{
Method or helper function for fit object of class \code{sfm}.
}
\author{
Eric Zivot and Sangeetha Srinivasan
}
\seealso{
\code{\link{fitSFM}}
}

