% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitted.tsfm}
\alias{fitted.tsfm}
\title{Get fitted values from a time series factor model}
\usage{
\method{fitted}{tsfm}(object, ...)
}
\arguments{
\item{object}{a fit object of class \code{tsfm} which is returned by
\code{\link{fitTSFM}}}

\item{...}{other arguments passed}
}
\value{
\item{fitted.xts}{an N x T data object of fitted values}
where, N is the number of assets and T is the number of time periods.
}
\description{
Method or helper function for fit object of class \code{tsfm}.
}
\examples{
\dontrun{
data(managers.df)
fit <- fitTSFM(asset.names=colnames(managers.df[,(1:6)]),
               factor.names=colnames(managers.df[,7:9]),
               market.name="SP500.TR",
               data=data, fit.method="OLS", variable.selection="none",
               add.up.market=TRUE, add.market.sqd=TRUE)
fitted(fit)
}
}
\author{
Eric Zivot and Sangeetha Srinivasan
}
\seealso{
\code{\link{fitTSFM}}
}

