% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTsfmLagLeadBeta.r
\name{fitTsfmLagLeadBeta}
\alias{fitTsfmLagLeadBeta}
\title{Fit a lagged and lead Betas factor model using time series regression}
\usage{
fitTsfmLagLeadBeta(
  asset.names,
  mkt.name,
  rf.name = NULL,
  data = data,
  fit.method = c("LS", "DLS", "Robust"),
  LagLeadBeta = 1,
  LagOnly = FALSE,
  control = fitTsfm.control(...),
  ...
)
}
\arguments{
\item{asset.names}{vector containing names of assets, whose returns or 
excess returns are the dependent variable.}

\item{mkt.name}{name of the column for market returns. It 
is required for a lagged Betas factor model.}

\item{rf.name}{name of the column of risk free rate variable to calculate 
excess returns for all assets (in \code{asset.names}) and the market factor (in 
\code{mkt.name}).Default is NULL, and no action is taken.}

\item{data}{vector, matrix, data.frame, xts, timeSeries or zoo object
containing column(s) named in \code{asset.names}, \code{factor.names} and 
optionally, \code{mkt.name} and \code{rf.name}.}

\item{fit.method}{the estimation method, one of "LS", "DLS" or "Robust". 
See details. Default is "LS".}

\item{LagLeadBeta}{A integer number to specify numbers of lags (and leads when LagOnly is FALSE) 
of Betas to 
include in the model. The Default is 1.}

\item{LagOnly}{Flag variable to only include the lags (or have both lags and leads). 
The Default is FALSE (both lags and leads).}

\item{control}{list of control parameters. The default is constructed by 
the function \code{\link{fitTsfm.control}}. See the documentation for 
\code{\link{fitTsfm.control}} for details.}

\item{...}{arguments passed to \code{\link{fitTsfm.control}}}
}
\value{
\code{fitTsfmLagLeadBeta} also returns an object of class \code{"tsfm"} like 
\code{fitTsfm}. The generic function such as \code{print}, \code{plot}, \code{predict} 
and \code{summary} methods exist. Also, the generic accessor functions \code{coef}, 
\code{fitted}, \code{residuals} and  \code{fmCov} can be applied as well.

An object of class \code{"tsfm"} is a list containing the following 
components:
\item{asset.fit}{list of fitted objects for each asset. Each object is of 
class \code{lm} if \code{fit.method="LS" or "DLS"}, class \code{lmRob} if 
the \code{fit.method="Robust"}.}
\item{alpha}{length-N vector of estimated alphas.}
\item{beta}{N x (L+1) matrix of estimated betas.}
\item{r2}{length-N vector of R-squared values.}
\item{resid.sd}{length-N vector of residual standard deviations.}
\item{call}{the matched function call.}
\item{data}{xts data object containing the assets and factors.}
\item{asset.names}{asset.names as input.}
\item{fit.method}{fit.method as input.}
Where N is the number of assets, L is the number of lagged and lead market Betas and T is the 
number of time periods.
}
\description{
This is a wrapper function to fits a time series lagged Betas factor model for one 
or more asset returns or excess returns using time series regression. 
Users can choose between ordinary least squares-LS, discounted least 
squares-DLS (or) robust regression like \code{fitTsfm}.An object of class 
\code{"tsfm"} is returned.
}
\details{
The lagged and lead returns model estimates lagged and lead market Beta. Specifically, 
\deqn{r_t = \alpha + \beta_0 MKT_t + \beta^-_1 MKT_t-1 + \ldots + \beta^-_K+1 MKT_t-K 
+  \beta^+_1 MKT_t+1 + \ldots + \beta^+_K MKT_t+K \epsilon_t, t=1 \ldots T}
where \eqn{r_t} is the asset returns, and MKT is the market factor. It is usually 
needed for illiquid securities with stale prices. One can also report the sum of 
the lagged and lead Betas: 
\deqn{\beta = \beta_0 + \beta^+_1 + \beta^+_1 + \ldots + \beta^+_K + 
\beta^-_1 + \ldots + \beta^-_K }
}
\examples{
# load data from the database
library(PerformanceAnalytics)
data(managers)

# example: A lagged Beetas model with LS fit
fit <- fitTsfmLagLeadBeta(asset.names=colnames(managers[,(1:6)]),LagLeadBeta=2,LagOnly=TRUE,
                      mkt.name="SP500 TR",rf.name="US 3m TR",data=managers)
summary(fit)
fitted(fit)

}
\references{
Scholes, M. and Williams, J. T. (1977). Estimating betas from non-synchronous
data, Journal of Financial Economics, vol. 5, 1977, pp. 309-327
}
\seealso{
The original time series function \code{\link{fitTsfm}} and its generic functions
 application.
}
\author{
Yi-An Chen.
}
