% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fmCov.R
\name{fmCov}
\alias{fmCov}
\alias{fmCov.sfm}
\alias{fmCov.tsfm}
\title{Covariance Matrix for assets' returns from fitted factor model.}
\usage{
fmCov(object, ...)

\method{fmCov}{tsfm}(object, use = "pairwise.complete.obs", ...)

\method{fmCov}{sfm}(object, use = "pairwise.complete.obs", ...)
}
\arguments{
\item{object}{fit object of class \code{tsfm}, \code{sfm} or \code{ffm}.}

\item{...}{optional arguments passed to \code{\link[stats]{cov}}.}

\item{use}{an optional character string giving a method for computing
covariances in the presence of missing values. This must be (an
abbreviation of) one of the strings "everything", "all.obs",
"complete.obs", "na.or.complete", or "pairwise.complete.obs". Default is
"pairwise.complete.obs".}
}
\value{
The computed \code{N x N} covariance matrix for asset returns based
on the fitted factor model.
}
\description{
Computes the covariance matrix for assets' returns based on a
fitted factor model. This is a generic function with methods for classes
\code{tsfm}, \code{sfm} and \code{ffm}.
}
\details{
\code{R(i, t)}, the return on asset \code{i} at time \code{t},
is assumed to follow a factor model of the form, \cr \cr
\code{R(i,t) = alpha(i) + beta(i)*f(t) + e(i,t)}, \cr \cr
where, \code{alpha(i)} is the intercept, \code{f(t)} is a {K x 1} vector of
factor returns at time \code{t}, \code{beta(i)} is a \code{1 x K} vector of
factor exposures and the error terms \code{e(i,t)} are serially
uncorrelated across time and contemporaneously uncorrelated across assets
so that \code{e(i,t) ~ iid(0,sig(i)^2)}. Thus, the variance of asset
\code{i}'s return is given by \cr \cr
\code{var(R(i)) = beta(i)*cov(F)*tr(beta(i)) + sig(i)^2}. \cr \cr
And, the \code{N x N} covariance matrix of asset returns is \cr \cr
\code{var(R) = B*cov(F)*tr(B) + D}, \cr \cr
where, B is the \code{N x K} matrix of factor betas and \code{D} is a
diagonal matrix with \code{sig(i)^2} along the diagonal.

The method for computing covariance can be specified via the \dots
argument. Note that the default of \code{use="pairwise.complete.obs"} for
handling NAs restricts the method to "pearson".
}
\examples{
# Time Series Factor model
data(managers)
fit <- fitTsfm(asset.names=colnames(managers[, (1:6)]),
               factor.names=c("EDHEC.LS.EQ","SP500.TR"), data=managers)
fmCov(fit)

# Statistical Factor Model
data(StockReturns)
sfm.pca.fit <- fitSfm(r.M, k=2)
fmCov(sfm.pca.fit)

\dontrun{
# Fundamental Factor Model
data(stock)
# there are 447 assets
exposure.names <- c("BOOK2MARKET", "LOG.MARKETCAP")
beta.mat <- subset(stock, DATE=="2003-12-31")[, exposure.names]
beta.mat1 <- cbind(rep(1, 447), beta.mat1)
# FM return covariance
fit.fund <- fitFfm(exposure.names=c("BOOK2MARKET", "LOG.MARKETCAP"),
                   data=stock, returnsvar="RETURN", datevar="DATE",
                   assetvar="TICKER", wls=TRUE, regression="classic",
                   covariance="classic", full.resid.cov=FALSE)
ret.cov.fundm <- fmCov(beta.mat1,fit.fund$factor.cov$cov,fit.fund$resid.sd)
fit.fund$returns.cov$cov == ret.cov.fundm
}
}
\author{
Eric Zivot, Yi-An Chen and Sangeetha Srinivasan.
}
\references{
Zivot, E., & Jia-hui, W. A. N. G. (2006). Modeling Financial Time
Series with S-Plus Springer-Verlag.
}
\seealso{
\code{\link{fitTsfm}}, \code{\link{fitSfm}}, \code{\link{fitFfm}}

\code{\link[stats]{cov}} for more details on arguments \code{use} and
\code{method}.
}

