% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmmc.R
\name{fmmc.estimate.se}
\alias{fmmc.estimate.se}
\title{Main function to calculate the standard errror of the estimate}
\usage{
fmmc.estimate.se(fmmcObjs, fun = NULL, se = FALSE, nboot = 100,
  parallel = FALSE)
}
\arguments{
\item{fmmcObjs}{A list of fmmc objects computed using .fmmc.proc and containing
bootstrapped returns}

\item{fun}{A callback function where the first argument is returns and all the
other arguments are bounded to values}

\item{se}{A flag to indicate if standard error for the estimate must be calculated}

\item{nboot}{Number of bootstrap samples}

\item{parallel}{A flag to indicate if multiple cpu cores must be used}
}
\value{
returns the estimates and thier standard errors given fmmc objects
}
\description{
Main function to calculate the standard errror of the estimate
}
\details{
This method takes in a list of fmmc objects and a callback function to compute
an estimate. The first argument of the callback function must be the data 
bootstrapped using fmmc procedure. The remaining arguments can be suitably
bound to the parameters as needed. This function can also be used to calculate
the standard error using the se flag.
}
\author{
Rohit Arora
}

