% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmmc.R
\name{fmmc}
\alias{fmmc}
\title{Compute fmmc objects that can be used for calcuation of estimates and their
standard errors}
\usage{
fmmc(R, factors, parallel = FALSE, ...)
}
\arguments{
\item{R}{matrix of returns in xts format}

\item{factors}{matrix of factor returns in xts format}

\item{parallel}{flag to utilize multiplecores on the cpu. All cores are used.}

\item{...}{Arguments that must be passed to fitTsfm}
}
\value{
returns an list of fmmc objects
}
\description{
Compute fmmc objects that can be used for calcuation of estimates and their
standard errors
}
\details{
This method takes in data and factors as xts objects where multiple
time series with different starting dates are merged together. It then
computes FMMC objects as described in Jiang and Martin (2013)
}
\references{
Yindeng Jiang and Richard Doug Martin. Better Risk and Performance 
Estimates with Factor Model Monte Carlo. SSRN Electronic Journal, July 2013.
}
\author{
Rohit Arora
}
