% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFfmDT.R
\name{lagExposures}
\alias{lagExposures}
\title{lagExposures allows the user to lag exposures by one time period}
\usage{
lagExposures(specObj)
}
\arguments{
\item{specObj}{an ffm specification object of of class \code{"ffmSpec"}}
}
\value{
specObj an ffm spec Object that has been lagged
}
\description{
Function lag the style exposures in the exposure matrix 
 by one time period.
}
