% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-risk.R
\name{make_factor_star_cov}
\alias{make_factor_star_cov}
\title{Build augmented factor covariance matrix}
\usage{
make_factor_star_cov(factor_cov)
}
\arguments{
\item{factor_cov}{K x K factor covariance matrix (with or without colnames)}
}
\value{
(K+1) x (K+1) matrix with "Residuals" row/column appended
}
\description{
Extends a K x K factor covariance matrix to (K+1) x (K+1) by adding a
unit-variance residual pseudo-factor with zero covariance to all real
factors.
}
\keyword{internal}
