% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-risk.R
\name{make_resid_diag}
\alias{make_resid_diag}
\title{Build diagonal residual covariance matrix}
\usage{
make_resid_diag(resid_var)
}
\arguments{
\item{resid_var}{N-vector of residual variances (sigma_e^2)}
}
\value{
N x N diagonal matrix
}
\description{
Constructs D.e = diag(sigma_e^2) for the factor model covariance
decomposition: Cov(R) = B Cov(F) B' + D.e.
}
\keyword{internal}
