% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{managers.ffm}
\alias{managers.ffm}
\title{managers data for ffm}
\source{
TBA
}
\usage{
data("managers.ffm")
}
\description{
Hypothetical Alternative Asset Manager and Benchmark Data for Time Series Factor Model Fit
}
