% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{managers}
\alias{managers}
\title{time-series data}
\source{
TBA

TBA
}
\usage{
data("managers")

data("managers")
}
\description{
Hypothetical Alternative Asset Manager and Benchmark Data for Time Series Factor Model Fit


}
