% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeData.R
\name{mergeSPGMICRSP}
\alias{mergeSPGMICRSP}
\title{Merge data SPGMI scores with CRSP stocks.}
\usage{
mergeSPGMICRSP()
}
\description{
Merge data SPGMI scores with CRSP stocks.
}
\examples{
data(stocksCRSP)
data(scoresSPGMI)
dataSPGMICRSP = mergeSPGMICRSP()
}
\author{
Kirk Li.
}
