% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mktUS}
\alias{mktUS}
\title{US Market Returns}
\source{
WRDS
}
\usage{
data("mktUS")
}
\description{
Monthly returns including all distributions, on a value-weighted market portfolio of NYSE/AMEX/NASDAQ
}
