% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.sfm.r
\name{plot.sfm}
\alias{plot.sfm}
\title{Plots from a fitted statistical factor model}
\usage{
\method{plot}{sfm}(x, which.plot.group = NULL, factor.subset = 1:4,
  asset.subset = 1:5, n.top = 3, plot.single = FALSE, asset.name,
  which.plot.single = NULL, colorset = (1:12), legend.loc = "topleft",
  las = 1, VaR.method = "historical", cum.var = TRUE, eig.max = 0.9,
  loop = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{sfm} produced by \code{fitSfm}.}

\item{which.plot.group}{a number to indicate the type of group plot for
multiple assets. If \code{NULL} (default), the following menu appears: \cr
1 = Screeplot of eigenvalues, \cr
2 = Time series plot of estimated factors, \cr
3 = Estimated factor loadings, \cr
4 = Histogram of R-squared, \cr
5 = Histogram of residual volatility,\cr
6 = Factor model residual correlation \cr
7 = Factor model correlation,\cr
8 = Factor contribution to SD,\cr
9 = Factor contribution to ES,\cr
10 = Factor contribution to VaR, \cr
11 = Factor mimicking portfolio weights - top long and short positions in each factor, \cr
12 = Asset correlations - top long and short positions in each factor}

\item{factor.subset}{vector of names/indices of factors to show for group
plots. Default is 1:4.}

\item{asset.subset}{vector of names/indices of assets to show for group
plots. Default is 1:5.}

\item{n.top}{scalar; number of largest and smallest weights to display for
each factor mimicking portfolio. Default is 3.}

\item{plot.single}{logical; If \code{TRUE} plots the characteristics of an
individual asset's factor model. The type of plot is given by
\code{which.plot.single}. Default is \code{FALSE}.}

\item{asset.name}{name of the individual asset to be plotted. Is necessary
if \code{x} contains multiple asset fits and \code{plot.single=TRUE}.}

\item{which.plot.single}{a number to indicate the type of group plot for an
individual asset. If \code{NULL} (default), the following menu appears: \cr
 1 = Time series plot of actual and fitted asset returns,\cr
 2 = Time series plot of residuals with standard error bands, \cr
 3 = Time series plot of squared residuals, \cr
 4 = Time series plot of absolute residuals,\cr
 5 = SACF and PACF of residuals,\cr
 6 = SACF and PACF of squared residuals,\cr
 7 = SACF and PACF of absolute residuals,\cr
 8 = Histogram of residuals with normal curve overlayed,\cr
 9 = Normal qq-plot of residuals,\cr
 10 = CUSUM test-Recursive residuals,\cr
 11 = CUSUM test-LS residuals,\cr
 12 = Recursive estimates (RE) test of LS regression coefficients,\cr
 13 = Rolling estimates over a 24-period observation window}

\item{colorset}{color palette to use for all the plots. Default is
\code{c(1:12)}. The 1st element will be used for individual time series
plots or the 1st series plotted, the 2nd element for the 2nd object in the
plot and so on.}

\item{legend.loc}{places a legend into one of nine locations on the chart:
"bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright",
"right", or "center". Default is "bottomright". Use \code{legend.loc=NULL}
to suppress the legend.}

\item{las}{one of {0, 1, 2, 3} to set the direction of axis labels, same as
in \code{plot}. Default here is 1.}

\item{VaR.method}{a method for computing VaR; one of "modified", "gaussian",
"historical" or "kernel". VaR is computed using
\code{\link[PerformanceAnalytics]{VaR}}. Default is "historical".}

\item{cum.var}{logical; If \code{TRUE}, the cumulative fraction of the
variance is printed above each bar in the screeplot of eigenvalues. Default
is \code{TRUE}.}

\item{eig.max}{scalar in (0,1] for limiting the screeplot to factors that
explain a given percent of the variance. Default is 0.9.}

\item{loop}{logical to indicate if the plot menu should be repeated. Default
is \code{TRUE}.}

\item{...}{further arguments to be passed to other plotting functions.}
}
\description{
Generic \code{plot} method for object of class \code{sfm}.
Plots chosen characteristic(s) for one or more assets.
}
\details{
If the plot type argument is not specified, a menu prompts for user input
and the corresponding plot is output. And, the menu is repeated for
user convenience in plotting multiple characteristics. Selecting '0' from
the menu exits the current \code{plot.sfm} call. Alternately, setting
\code{loop=FALSE} will exit after plotting any one chosen characteristic
without the need for menu selection.

Group plots are the default. The variables in \code{asset.subset} and factors
in \code{asset.subset} are plotted depending on the characteristic chosen.
The default is to show the first 4 factors and first 5 assets.

Individual asset plots are selected by specifying \code{plot.single=TRUE}.
In which case, \code{asset.name} is necessary if multiple assets
were modeled in \code{x}. However, if the \code{fitSfm} object contains only
one asset's factor model fit, \code{plot.sfm} can infer this automatically,
without user input.
}
\examples{
# load data from the database
data(StockReturns)

# APCA with number of factors, k=15
fit.apca <- fitSfm(r.W, k=15, refine=TRUE)

# group plot(default); select type from menu prompt
# menu is auto-looped to get multiple types of plots based on the same fit
# plot(fit.apca)

# plot the factor betas of 1st 4 assets fitted above
# loop disabled to get one type of plot without interative menu
plot(fit.apca, asset.subset=1:4, which.plot.group=3, loop=FALSE)

# plot factor model return correlation; angular order of the eigenvectors
plot(fit.apca, which.plot.group=7, loop=FALSE,
     order="AOE", method="ellipse", tl.pos = "d")

# histogram of residuals from an individual asset's factor model fit
plot(fit.apca, plot.single=TRUE, asset.name="AFL", which.plot.single=8,
     loop=FALSE)
}
\author{
Eric Zivot, Sangeetha Srinivasan and Yi-An Chen
}
\seealso{
\code{\link{fitSfm}} and \code{\link{summary.sfm}} for details
about the time series factor model fit, extractor functions and summary
statistics.

\code{\link[strucchange]{efp}} for CUSUM tests.

\code{\link[xts]{plot.xts}},
\code{\link[PerformanceAnalytics]{chart.TimeSeries}},
\code{\link[PerformanceAnalytics]{chart.ACFplus}},
\code{\link[PerformanceAnalytics]{chart.Histogram}},
\code{\link[PerformanceAnalytics]{chart.QQPlot}},
\code{\link[graphics]{barplot}}, \code{\link[lattice]{barchart}} and
\code{\link[corrplot]{corrplot}} for plotting methods used.

\code{\link{fmSdDecomp}}, \code{\link{fmEsDecomp}},
\code{\link{fmVaRDecomp}} for factor model risk measures.
}

