% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.tsfm}
\alias{plot.tsfm}
\title{Plots from a fitted time series factor model}
\usage{
\method{plot}{tsfm}(x, colorset = c(1:12), legend.loc = NULL,
  which.plot = NULL, max.show = 6, plot.single = FALSE, asset.name,
  which.plot.single = NULL, VaR.method = "historical", ...)
}
\arguments{
\item{x}{an object of class \code{tsfm} produced by \code{fitTsfm}.}

\item{colorset}{a vector of colors for the bars or bar components. Argument
is used by \code{\link[graphics]{barplot}}. Default is c(1:12).}

\item{legend.loc}{places a legend into one of nine locations on the chart:
bottomright, bottom, bottomleft, left, topleft, top, topright, right, or
center. Argument is used by
\code{\link[PerformanceAnalytics]{chart.TimeSeries}}. Default is \code{NULL}.}

\item{which.plot}{a number to indicate the type of group plot for multiple
assets. Default is NULL; which brings up the following menu: \cr
1 = "Actual and Fitted asset returns", \cr
2 = "R-squared", \cr
3 = "Residual Volatility",\cr
4 = "Factor Model Correlation",\cr
5 = "Factors' Contribution to SD",\cr
6 = "Factors' Contribution to ES",\cr
7 = "Factors' Contribution to VaR"}

\item{max.show}{maximum number of assets in a plot. Default is 6.}

\item{plot.single}{a logical value. If \code{TRUE}, plots an individual
asset's linear factor model trait selected by \code{which.plot.single}.
Default is \code{FALSE}.}

\item{asset.name}{name of the individual asset to be plotted. Is necessary
if \code{plot.single=TRUE}}

\item{which.plot.single}{a number to indicate the type of group plot for an
individual asset. Default is NULL; which brings up the following menu: \cr
 1 = Time series plot of actual and fitted factor returns,\cr
 2 = Time series plot of residuals with standard error bands, \cr
 3 = Time series plot of squared residuals, \cr
 4 = Time series plot of absolute residuals,\cr
 5 = SACF and PACF of residuals,\cr
 6 = SACF and PACF of squared residuals,\cr
 7 = SACF and PACF of absolute residuals,\cr
 8 = Histogram of residuals with normal curve overlayed,\cr
 9 = Normal qq-plot of residuals,\cr
 10= CUSUM plot of recursive residuals,\cr
 11= CUSUM plot of OLS residuals,\cr
 12= CUSUM plot of recursive estimates relative to full sample estimates,\cr
 13= Rolling estimates over a 24-period observation window}

\item{VaR.method}{a method for computing VaR; one of "modified", "gaussian",
"historical" or "kernel". VaR is computed using
\code{\link[PerformanceAnalytics]{VaR}}. Default is "historical".}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generic \code{plot} method for object of class \code{tsfm}.
Plots chosen characteristic(s) for one or more assets.
}
\examples{
\dontrun{
# load data from the database
data(managers)
fit.macro <- fitTsfm(asset.names=colnames(managers[,(1:6)]),
                     factor.names=colnames(managers[,(7:8)]),
                     rf.name="US 3m TR", data=managers)
# plot all assets and show only the first 4 assets.
plot(fit.macro, max.show=4)
# plot of an individual asset, "HAM1"
plot(fit.macro, plot.single=TRUE, asset.name="HAM1")
}
}
\author{
Eric Zivot, Yi-An Chen and Sangeetha Srinivasan
}
\seealso{
\code{\link{fitTsfm}}, \code{\link{summary.tsfm}}
}

