% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portEsDecomp.R
\name{portEsDecomp}
\alias{portEsDecomp}
\alias{portEsDecomp.tsfm}
\title{Decompose portfolio ES into individual factor contributions}
\usage{
portEsDecomp(object, ...)

\method{portEsDecomp}{tsfm}(object, weights = NULL, p = 0.95,
  type = c("np", "normal"), use = "pairwise.complete.obs", ...)
}
\arguments{
\item{object}{fit object of class \code{tsfm}, or \code{ffm}.}

\item{...}{other optional arguments passed to \code{\link[stats]{quantile}} and 
optional arguments passed to \code{\link[stats]{cov}}}

\item{weights}{a vector of weights of the assets in the portfolio. Default is NULL.}

\item{p}{confidence level for calculation. Default is 0.95.}

\item{type}{one of "np" (non-parametric) or "normal" for calculating VaR. 
Default is "np".}

\item{use}{an optional character string giving a method for computing factor
covariances in the presence of missing values. This must be (an 
abbreviation of) one of the strings "everything", "all.obs", 
"complete.obs", "na.or.complete", or "pairwise.complete.obs". Default is 
"pairwise.complete.obs".}
}
\value{
A list containing 
\item{ES.fm}{length-1 vector of factor model ES of portfolio returns.}
\item{n.exceed}{length-1 vector of number of observations beyond VaR for 
portfolio.}
\item{idx.exceed}{numeric vector of index values of exceedances.}
\item{mES}{length-(N + K) vector of marginal contributions to VaR.}
\item{cES}{length-(N + K) vector of component contributions to VaR.}
\item{pcES}{length-(N + K) vector of percentage component contributions to VaR.}
Where, K is the number of factors and N is the number of assets.
}
\description{
Compute the factor contributions to Expected Tail Loss or 
Expected Shortfall (ES) of portfolio returns  based on Euler's theorem, given 
the fitted factor model. The partial derivative of ES with respect to factor 
beta is computed as the expected factor return given fund return is less 
than or equal to its value-at-risk (VaR). Option to choose between 
non-parametric and Normal.
}
\details{
The factor model for an asset's return at time \code{t} has the 
form \cr \cr \code{R(t) = beta'f(t) + e(t) = beta.star'f.star(t)} \cr \cr 
where, \code{beta.star=(beta,sig.e)} and \code{f.star(t)=[f(t)',z(t)]'}. By 
Euler's theorem, the ES of the asset's return is given by:
\cr \cr \code{ES.fm = sum(cES_k) = sum(beta.star_k*mES_k)} \cr \cr
where, summation is across the \code{K} factors and the residual, 
\code{cES} and \code{mES} are the component and marginal 
contributions to \code{ES} respectively. The marginal contribution to ES is
defined as the expected value of \code{F.star}, conditional on the loss 
being less than or equal to \code{VaR.fm}. This is estimated as a sample 
average of the observations in that data window.
}
\examples{
# Time Series Factor Model
require(factorAnalytics)
data(managers)
fit.macro <- fitTsfm(asset.names=colnames(managers[,(1:6)]),
                     factor.names=colnames(managers[,(7:8)]), data=managers)
ES.decomp <- portEsDecomp(fit.macro)
# get the component contributions
ES.decomp$cES

}
\author{
Douglas Martin, Lingjie Yi
}

