% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portRsq.R
\name{portRsq}
\alias{portRsq}
\title{Portfolio R-squared and Adjusted R-squared}
\usage{
portRsq(z, weight = NULL, ...)
}
\arguments{
\item{z}{an object of class \code{ffm} produced by \code{fitFfm}}

\item{weight}{a vector of weights of the assets in the portfolio. Default is NULL.}

\item{...}{additional arguments unused}
}
\description{
Calcluate the R-squared and Adjusted R-squared for the portfolio of stocks
}
\examples{

#Load the data 
data("stocks145scores6")
 
#Fit a Ffm
fit <- fitFfm(data=data145, asset.var="TICKER", ret.var="RETURN", 
              date.var="DATE", exposure.vars="SECTOR")
              
#Find the portfolio R-squared and adjusted portfolio R-squared with default weights.               
portRsq(fit)

}
\author{
Avinash Acharya
}

