\name{predict.FundamentalFactorModel}
\alias{predict.FundamentalFactorModel}
\title{predict method for FundamentalFactorModel object}
\usage{
  predict.FundamentalFactorModel(fit, newdata)
}
\arguments{
  \item{fit}{"FundamentalFactorModel" object}
}
\description{
  Generic function of predict method for
  fitFundamentalFactorModel.
}
\author{
  Yi-An Chen
}

