% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sfm.r
\name{predict.sfm}
\alias{predict.sfm}
\title{Predicts asset returns based on a fitted statistical factor model}
\usage{
\method{predict}{sfm}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{sfm} produced by \code{fitSfm}.}

\item{newdata}{a vector, matrix, data.frame, xts, timeSeries or zoo object 
containing the variables with which to predict.}

\item{...}{optional arguments passed to \code{predict.lm}.}
}
\value{
\code{predict.sfm} produces a vector or a matrix of predictions.
}
\description{
S3 \code{predict} method for object of class \code{sfm}. It 
calls the \code{predict} method for fitted objects of class \code{lm}.
}
\examples{
# load data from the database
data(StockReturns)
# fit the factor model with PCA
fit <- fitSfm(r.M, k=2)

pred.fit <- predict(fit)
newdata <- data.frame("CITCRP"=rnorm(n=120), "CONED"=rnorm(n=120))
rownames(newdata) <- rownames(fit$data)
pred.fit2 <- predict(fit, newdata, interval="confidence")

}
\seealso{
\code{\link{fitSfm}}, \code{\link{summary.sfm}}
}
\author{
Yi-An Chen and Sangeetha Srinivasan
}
