% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predict.tsfm.r
\name{predict.tsfm}
\alias{predict.tsfm}
\title{Predicts asset returns based on a fitted time series factor model}
\usage{
\method{predict}{tsfm}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{tsfm} produced by \code{fitTsfm}.}

\item{newdata}{a vector, matrix, data.frame, xts, timeSeries or zoo object
containing the variables with which to predict.}

\item{...}{optional arguments passed to \code{predict.lm} or
\code{\link[robust]{predict.lmRob}}, such as \code{se.fit}, or, to
\code{\link[lars]{predict.lars}} such as \code{mode}.}
}
\value{
\code{predict.tsfm} produces a vector or a matrix of predictions.
}
\description{
S3 \code{predict} method for object of class \code{tsfm}. It
calls the \code{predict} method for fitted objects of class \code{lm},
\code{lmRob} or \code{lars} as appropriate.
}
\examples{
# load data from the database
data(managers)
# fit the factor model with OLS
fit <- fitTsfm(asset.names=colnames(managers[,(1:6)]),
               factor.names=c("EDHEC LS EQ","SP500 TR"), data=managers)

pred.fit <- predict(fit)
newdata <- data.frame("EDHEC LS EQ"=rnorm(n=120), "SP500 TR"=rnorm(n=120))
rownames(newdata) <- rownames(fit$data)
pred.fit2 <- predict(fit, newdata, interval="confidence")
}
\author{
Yi-An Chen and Sangeetha Srinivasan
}
\seealso{
\code{\link{fitTsfm}}, \code{\link{summary.tsfm}}
}

