% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.pafm}
\alias{print.pafm}
\title{Print object of class \code{"pafm"}.}
\usage{
\method{print}{pafm}(x, ...)
}
\arguments{
\item{x}{object of class \code{"pafm"} created by
\code{paFM}.}

\item{...}{Other arguments for \code{print} methods.}
}
\description{
Generic function of print method for \code{paFM}.
}
\examples{
# load data from the database
 data(managers.df)
# fit the factor model with OLS
fit <- fitTSFM(asset.names=colnames(managers.df[,(1:6)]),
               factor.names=c("EDHEC.LS.EQ","SP500.TR"), data=managers.df,
               add.up.market=FALSE, add.market.sqd=FALSE,
               fit.method="OLS", variable.selection="none")
fm.attr <- paFM(fit)
print(fm.attr)
}
\author{
Yi-An Chen.
}

