% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sfm.r
\name{print.sfm}
\alias{print.sfm}
\title{Prints a fitted statistical factor model}
\usage{
\method{print}{sfm}(x, digits = max(3, .Options$digits - 3), ...)
}
\arguments{
\item{x}{an object of class \code{sfm} produced by \code{fitSfm}.}

\item{digits}{an integer value, to indicate the required number of 
significant digits. Default is 3.}

\item{...}{optional arguments passed to the \code{print} method.}
}
\description{
S3 \code{print} method for object of class \code{sfm}. Prints 
the call, factor model dimensions and summary statistics for the estimated 
factor loadings, r-squared values and residual volatilities from the fitted 
object.
}
\examples{
data(StockReturns)
fit <- fitSfm(r.M, k=2)
print(fit)

}
\seealso{
\code{\link{fitSfm}}, \code{\link{summary.sfm}}
}
\author{
Yi-An Chen and Sangeetha Srinivasan
}
