% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.summary.tsfm}
\alias{print.summary.tsfm}
\title{Print method for object of class "summary.tsfm"}
\usage{
\method{print}{summary.tsfm}(object, digits = 3, ...)
}
\arguments{
\item{object}{an object of class \code{summary.tsfm}.}

\item{digits}{number of significants digits to use when printing.}

\item{...}{futher arguments passed to or from other methods.}
}
\description{
S3 \code{summary} method for object of class {summary.tsfm}.
}

