% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.tsfm}
\alias{print.tsfm}
\title{Prints out a fitted time series factor model object}
\usage{
print.tsfm(x, digits = max(3, .Options$digits - 3), ...)
}
\arguments{
\item{x}{an object of class \code{tsfm} produced by \code{fitTSFM}.}

\item{digits}{an integer value, to indicate the required number of
significant digits. Default is 3.}

\item{...}{optional arguments passed to the \code{print} method.}
}
\description{
S3 \code{print} method for object of class \code{tsfm}. Prints
the call, factor model dimension, regression coefficients, r-squared and
residual volatilities from the fitted object.
}
\examples{
data(managers.df)
fit <- fitTSFM(asset.names=colnames(managers.df[,(1:6)]),
               factor.names=colnames(managers.df[,7:9]),
               market.name="SP500.TR",
               data=data, fit.method="OLS", variable.selection="none",
               add.up.market=TRUE, add.market.sqd=TRUE)
print(fit)
}
\author{
Yi-An Chen and Sangeetha Srinivasan
}
\seealso{
\code{\link{fitTSFM}}, \code{\link{summary.tsfm}},
\code{\link{tsfm}}
}

