% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadDataGitHub.R
\name{readCRSPdataGithub}
\alias{readCRSPdataGithub}
\title{download CRSP dataset from github}
\usage{
readCRSPdataGithub(filename = "")
}
\arguments{
\item{filename}{stockCRSPdaily, stockCRSPweekly, stockCRSPmonthly}
}
\value{

}
\description{
download CRSP dataset from github
}
\examples{
readCRSPdataGithub("stockCRSPweekly")

}
\author{
Kirk Li.
}
