% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repReturn.R
\name{repReturn.ffm}
\alias{repReturn.ffm}
\title{Portfolio return reports for risk decomposition and performance analysis}
\usage{
\method{repReturn}{ffm}(ffmObj, weights = NULL, isPlot = FALSE, ...)
}
\arguments{
\item{ffmObj}{fit object of class \code{tsfm}, \code{sfm} or \code{ffm}.}

\item{isPlot}{logical variable to generate plot or not.}

\item{...}{additional arguments unused}

\item{weight}{a vector of weights of the assets in the portfolio. Default is NULL.}
}
\description{
conduct portfolio return analysis reporting
}
\examples{

#Load the data 
data("stocks145scores6")
 
#Fit a Ffm
fit <- fitFfm(data = data145, # Change fit object to mixed.mod
exposure.vars = c("SECTOR","ROE","BP","PM12M1M","SIZE","ANNVOL1M","EP"),
date.var = "DATE", 
ret.var = "RETURN", 
asset.var = "TICKER", 
fit.method="WLS",
z.score = F)
              
Conduct portfolio returns analysis reporting with default weights.               
repReturn(fit)

}
\author{
Lingjie Yi
}

