% Generated by roxygen2 (4.0.1): do not edit by hand
\name{residuals.sfm}
\alias{residuals.sfm}
\title{Get residuals from a fitted stochastic factor model}
\usage{
\method{residuals}{sfm}(x)
}
\arguments{
\item{x}{an object of class \code{sfm} which is returned by
\code{\link{fitSFM}}}
}
\value{
\item{residuals.xts}{an N x T data object of residuals}
where, N is the number of assets and T is the number of time periods.
}
\description{
Method or helper function for fit object of class \code{sfm}.
}
\author{
Eric Zivot and Sangeetha Srinivasan
}
\seealso{
\code{\link{fitSFM}}
}

