% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsqTstatsTs.R
\name{rsqTstatsTs}
\alias{rsqTstatsTs}
\title{t-stats and Plots for a fitted Fundamental Factor Model}
\usage{
rsqTstatsTs(ffmObj, isPlot = TRUE, col = "cyan", z.alpha = 1.96,
  layout = c(2, 3), type = "h", ...)
}
\arguments{
\item{ffmObj}{an object of class \code{ffm} produced by \code{fitFfm}}

\item{isPlot}{logical. If \code{TRUE} the barplots are plotted.}

\item{col}{specification for the default plotting color. Default is cyan}

\item{z.alpha}{critical value corresponding to the confidence interval. Default is 1.96 i.e 95\% C.I}

\item{layout}{numeric vector of length 2 or 3 giving the number of columns, rows, and pages (optional) in the xyplot of t-statistics. Default is c(2,3).}

\item{type}{character. Type of the xyplot of t-statistics; \code{"l"} for lines, \code{"p"} for points, \code{"h"} for histogram like (or high-density) vertical lines
and \code{"b"} for both. Deafault is \code{"h"}.}

\item{...}{potentially further arguments passed.}
}
\value{
\code{rsqTstatsTs} plots the R-squared, t-stats and significant t-stats values  if \code{isPlot} is \code{TRUE} and returns a list with following components:
\item{R-squared}{ length-T vector of R-squared values.}
\item{tstats}{ an xts object of t-stats values.}
\item{z.alpha}{ critical value corresponding to the confidence interval.}
}
\description{
Calculate and plot the time series of the R-squared values, t-statistic vales and the
 number of risk indices with significant t-stats for a fundamentally fit object.
}
\examples{
 
 data("factorDataSetDjia5Yrs")
 
#Fit a Ffm
 require(factorAnalytics)
 fit <- fitFfm(data = factorDataSetDjia5Yrs,exposure.vars = c("MARKETCAP","ENTVALUE","P2B","EV2S"),
               date.var = "DATE", ret.var = "RETURN", asset.var = "TICKER", fit.method="WLS")
               
 #Find r2, tstats  with C.I = 95\% and plot the results.
 stats = rsqTstatsTs(fit, isPlot = TRUE, col = "blue", z.alpha =1.96)  
   
              

}
\author{
Doug Martin, Avinash Acharya
}

