% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{scoresSPGMI}
\alias{scoresSPGMI}
\title{cleaned SPGMI data}
\source{
S&P Global Market Intelligence
}
\usage{
data("scoresSPGMI")
}
\description{
cleaned 300 stocks' 12 factor scores data
}
