% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{scoresSPGMIraw}
\alias{scoresSPGMIraw}
\title{raw SPGMI data}
\source{
TBA

TBA
}
\usage{
data("scoresSPGMIraw")

data("scoresSPGMIraw")
}
\description{
raw 300 stocks' 12 factor scores data

raw 300 stocks' 12 factor scores data
}
