% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCRSPandSPGMI.R
\name{selectCRSPandSPGMI}
\alias{selectCRSPandSPGMI}
\title{Select CRSP Stocks and SPGMI Factors}
\usage{
selectCRSPandSPGMI(
  stocks = stocksCRSP,
  factors = factorsSPGMI,
  dateSet = c("2006-01-31", "2010-12-31"),
  stockItems = c("Date", "TickerLast", "CapGroup", "Sector", "Return", "Ret13WkBill",
    "mktIndexCRSP"),
  factorItems = c("BP", "LogMktCap", "SEV"),
  capChoice = "SmallCap",
  Nstocks = 20
)
}
\arguments{
\item{stocks}{xxx}

\item{factors}{xxx}

\item{dateSet}{xxx}

\item{stockItems}{xxx}

\item{factorItems}{xxx}

\item{capChoice}{xxx}

\item{Nstocks}{xxx}
}
\value{
A merged data.frame consisting of selected stocks and factors
}
\description{
Select and merge a set of CRSP stocks and SPGMI factors and
SPGMI factors over a range of dates between January 1993 and December 2015
}
\details{
The date range is specified by a beginning Date object and an ending Date 
object.  Default values for the stocks and factors component names,
marketcap group, and number of stocks in the marketcap group can be
changed by the user
}
\examples{
data(stocksCRSP)
data(factorsSPGMI)

stocks_factors <- selectCRSPandSPGMI()

}
