% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateARL.R
\name{simulateARL}
\alias{simulateARL}
\title{Simulation for thresholds of the Lindley's recursion}
\usage{
simulateARL(mu, Threshold, delta, k = 3, EW_constant = 0.9,
  Fixed_Sigma = 1)
}
\arguments{
\item{mu}{A numeric value that determines the information ratio we want to
simulate thresholds for. No default value is set.}

\item{Threshold}{A numeric value that determines the threshold for the Lindley's
recursion to be updated in the recursion. No default value is set.}

\item{delta}{A numeric value representing the simulation accuracy.
In other words, we will simulate until k * sigma / mu < precision.}

\item{k}{A numeric value that determines the level of cut-off. Default is 3.}

\item{EW_constant}{A numeric value representing the ratio between the former sigma
and the new sigma. The default is set to 0.9}

\item{Fixed_Sigma}{The logical value representing if the sigma should be constant of not.
0 represents fixed sigma, 1 represents weighted sigma. The default is set to 1.}
}
\value{
\code{simulateARL} returns a \code{vector} of the following:
\item{ARL}{The average return length}
\item{s}{The standard deviation of the ARLs}
}
\description{
This function simulates thresholds of Lindley's recursion
used in the function \code{cusumActMgr}.
}
\details{
Simulate the thresholds used in \code{cusumActMgr}
}
\examples{
Lower_Threshold = 1.00
Upper_Threshold = 11.0
#Monthly mu's for monthly sigma=1, annualized IR = +0.5, 0
mu = c(0.5, 0) / sqrt(12)
Seq_M = 30:40
Thresholds = Lower_Threshold + Seq_M * (Upper_Threshold - Lower_Threshold) / 100
Threshold_upper = sapply(Thresholds, FUN = simulateARL, mu = mu[1], delta = 0.05)
Threshold_lower = sapply(Thresholds, FUN = simulateARL, mu = mu[2], delta = 0.05)

}
\references{
Philips, T. K., Yashchin, E. and Stein, D. M. (2003). "Using Statistical
Process Control to Monitor Active Managers", Journal of Portfolio Management,
Fall 2003, pp. 86-94.
}
\author{
Thomas Philips, Chindhanai Uthaisaad.
}
