% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFfmDT.R
\name{specFfm}
\alias{specFfm}
\title{Specifies the elements of a fundamental factor model}
\usage{
specFfm(
  data,
  asset.var,
  ret.var,
  date.var,
  exposure.vars,
  weight.var = NULL,
  addIntercept = FALSE,
  rob.stats = FALSE
)
}
\arguments{
\item{data}{data.frame of the balanced panel data containing the variables
\code{asset.var}, \code{ret.var}, \code{exposure.vars}, \code{date.var} and
optionally, \code{weight.var}.}

\item{asset.var}{character; name of the variable  for asset names.}

\item{ret.var}{character; name of the variable for asset returns.}

\item{date.var}{character; name of the variable containing the dates
coercible to class \code{Date}.}

\item{exposure.vars}{vector; names of the variables containing the
fundamental factor exposures.}

\item{weight.var}{character; name of the variable containing the weights
used when standarizing style factor exposures. Default is \code{NULL}. See
Details.}

\item{addIntercept}{logical; If \code{TRUE}, intercept is added in
the exposure matrix. Default is \code{FALSE},}

\item{rob.stats}{logical; If \code{TRUE}, robust estimates of covariance,
correlation, location and univariate scale are computed as appropriate (see
Details). Default is \code{FALSE}.}
}
\description{
Factor models have a few parameters that describe how the 
fitting is done.  This function summarizes them and returns a spec object for
cross-sectional regressions.  It also preps the data. An object of class 
\code{"ffmSpec"} is returned.
}
