\name{stat.fm.data}
\alias{sfm.dat}
\alias{sfm.apca.dat}
\docType{data}
\title{
Monthly Stock Return Data || Portfolio of Weekly Stock Returns
}
\description{
sfm.dat:
This is a monthly "data.frame" object from January 1978 to December 1987, with seventeen columns representing monthly returns of certain assets, as in Chapter 2 of Berndt (1991). 
sfm.apca.dat: 
This is a weekly "data.frame" object with dimension 182 x 1618, which runs from January 8, 1997 to June 28, 2000 and represents the stock returns on 1618 U.S. stocks.
}
\usage{data(stat.fm.data)}
\details{
CITCRP 
monthly returns of Citicorp. 
CONED 
monthly returns of Consolidated Edison. 
CONTIL 
monthly returns of Continental Illinois. 
DATGEN 
monthly returns of Data General. 
DEC 
monthly returns of Digital Equipment Company. 
DELTA 
monthly returns of Delta Airlines. 
GENMIL 
monthly returns of General Mills. 
GERBER 
monthly returns of Gerber. 
IBM 
monthly returns of International Business Machines. 
MARKET 
a value-weighted composite monthly returns based on transactions from the New York Stock Exchange and the American Exchange. 
MOBIL 
monthly returns of Mobile. 
PANAM 
monthly returns of Pan American Airways. 
PSNH 
monthly returns of Public Service of New Hampshire. 
TANDY 
monthly returns of Tandy. 
TEXACO 
monthly returns of Texaco. 
WEYER 
monthly returns of Weyerhauser. 
RKFREE 
monthly returns on 30-day U.S. Treasury bills. 
}
\source{
S+FinMetrics Berndt.dat & folio.dat
}
\references{
Berndt, E. R. (1991). The Practice of Econometrics: Classic and Contemporary. Addison-Wesley Publishing Co. 
}
\keyword{datasets}
