\name{stock}
\alias{stock}
\docType{data}
\title{
constructed NYSE 447 assets from 1996-01-01 through 2003-12-31.
}
\description{
constructed NYSE 447 assets from 1996-01-01 through 2003-12-31.
}
\details{
Continuous data: 
PRICE, RETURN, VOLUME, SHARES.OUT, MARKET.EQUITY,LTDEBT, NET.SALES, COMMON.EQUITY, NET.INCOME, STOCKHOLDERS.EQUITY, LOG.MARKETCAP, LOG.PRICE, BOOK2MARKET
Categorical data: 
GICS, GICS.INDUSTRY, GICS.SECTOR
Identication data:
DATE, PERMNO, TICKER.x
}
\references{
Guy Yullen and Yi-An Chen
}
\keyword{datasets}
