% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{stocks145scores6}
\alias{stocks145scores6}
\title{Factor Data}
\source{
TBA
}
\usage{
data(stocks145scores6)
}
\description{
Contains returns for 145 stocks spanned across 10 Sectors along with 6 factors: ROE, BP, PM12M1M, EP, ANNVOL1M, SIZE)
             starting from Jan 2000 to march 2013.
}

