% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{stocksCRSPscoresSPGMI}
\alias{stocksCRSPscoresSPGMI}
\title{cleaned SPGMI and CRSP data}
\source{
TBA
}
\usage{
data("stocksCRSPscoresSPGMI")
}
\description{
cleaned 300 stocks' 12 factor scores that merged with CRSP data
}
