% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{stocksCRSPscoresSPGMIraw}
\alias{stocksCRSPscoresSPGMIraw}
\title{raw SPGMI and CRSP data}
\source{
TBA
}
\usage{
data("stocksCRSPscoresSPGMIraw")
}
\description{
raw 300 stocks' 12 factor scores that merged with CRSP data
}
