% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{stocksWithFactorsData300}
\alias{stocksWithFactorsData300}
\title{300 stocks with factors}
\source{
TBA
}
\usage{
data("stocksWithFactorsData300")
}
\description{
the dataset has been replaced with stocksCRSPscoresSPGM
}
