% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repCusum.R
\name{summary.cusumActMgr}
\alias{summary.cusumActMgr}
\alias{print.summary.cusumActMgr}
\title{Summarizing a cusumActMgr object}
\usage{
\method{summary}{cusumActMgr}(object, digits = 3, ...)

\method{print}{summary.cusumActMgr}(x, digits = 3, ...)
}
\arguments{
\item{object}{an object of class \code{cusumActMgr} returned from \code{cusumActMgr}.}

\item{digits}{number of significants digits to use when printing.
Default is 3.}

\item{...}{futher arguments passed to or from other methods.}

\item{x}{an object of class \code{summary.cusumActMgr} returned from
\code{summary.cusumActMgr}.}
}
\value{
Returns an object of class \code{summary.ffm}.

Object of class \code{summary.cusumActMgr} is a list of length 10 containing:
\item{Logarithmic Excess Returns}{The logarithmic excess returns of the fund
relative to the benchmark}
\item{Annualized Moving Average}{The annualized moving average of the
logarithmic excess returns}
\item{Tracking Error}{The monthly tracking error of the logarithmic excess returns}
\item{Information Ratios}{The vector of monthly information ratios}
\item{Lindley's Recursion}{The vector  Lindley's recursion with a reset after the detection threshold (6.81) is passed.}
\item{Annualized Cusum IR}{The vector annualized CUSUM of the information ratios}
\item{Annualized Cusum Excess Return}{The vector annualized CUSUM of the excess returns}
\item{Excess Volatility}{Excess volatility of the fund, the benchmark and the excess return}
\item{Summary Annualized cusumIR}{The summary of annualized cusum IR}
\item{Summary Annualized Cusum Excess Returns}{The summary of annualized cusum excess returns}
}
\description{
\code{summary} method for object of class \code{cusumActMgr}.
Returned object is of class {summary.cusumActMgr}. The resulting object is fed to
\code{print.summary.cusumActMgr} to print all the summarized objects.
}
\examples{
data(cusumData)
results = cusumActMgr(portfolioName = "Parvest", benchmarkName = "RUS2500", data = cusumData)
x = summary(results)
print(x)

}
\author{
Chindhanai Uthaisaad.
}
