\name{summary.FundamentalFactorModel}
\alias{summary.FundamentalFactorModel}
\title{summary method for FundamentalFactorModel}
\usage{
  summary.FundamentalFactorModel(fit)
}
\arguments{
  \item{fit}{it object created by
  fitFundamentalFactorModel.}
}
\description{
  Generic function of summary method for
  fitTimeSeriesFactorModel.
}
\author{
  Yi-An Chen
}

