% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.sfm}
\alias{print.summary.sfm}
\alias{summary.sfm}
\title{Summarizing a fitted time series factor model}
\usage{
\method{summary}{sfm}(object, se.type = "Default", ...)

\method{print}{summary.sfm}(x, digits = 3, ...)
}
\arguments{
\item{object}{an object of class \code{sfm} returned by \code{fitSfm}.}

\item{se.type}{one of "Default", "HC" or "HAC"; option for computing HC/HAC
standard errors and t-statistics.}

\item{...}{futher arguments passed to or from other methods.}

\item{x}{an object of class \code{summary.sfm}.}

\item{digits}{number of significants digits to use when printing.
Default is 3.}
}
\value{
Returns an object of class \code{summary.sfm}.
The print method for class \code{summary.sfm} outputs the call,
coefficients (with standard errors and t-statistics), r-squared and
residual volatilty (under the homoskedasticity assumption) for all assets.

Object of class \code{summary.sfm} is a list of length N+2 containing:
\item{call}{the function call to \code{fitSfm}}
\item{se.type}{standard error type as input}
\item{sum.list}{list of summaries for the N fit objects of class \code{lm}
for each asset in the factor model.}
\item{mimic}{N x K matrix of factor mimicking portfolio weights.}
}
\description{
\code{summary} method for object of class \code{sfm}.
Returned object is of class {summary.sfm}.
}
\details{
The default \code{summary} method for a fitted \code{lm} object
computes the standard errors and t-statistics under the assumption of
homoskedasticty. Argument \code{se.type} gives the option to compute
heteroskedasticity-consistent (HC) or
heteroskedasticity-autocorrelation-consistent (HAC) standard errors and
t-statistics using \code{\link[lmtest]{coeftest}}.
}
\examples{
data(StockReturns)
# fit the factor model with PCA
fit <- fitSfm(r.M, k=2)

# summary of factor model fit for all assets
summary(fit, "HAC")
}
\author{
Sangeetha Srinivasan
}
\seealso{
\code{\link{fitSfm}}, \code{\link[stats]{summary.lm}}
}

