% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wtsStocks145Gmv}
\alias{wtsStocks145Gmv}
\title{CRSP 145 stocks GMV portfolio weights}
\source{
TBA
}
\usage{
data("wtsStocks145Gmv")
}
\description{
Contains weights obtained after optimizing the portfolio returns of 145 stocks (from dataset stocks145scores6) 
             for a global minimum variance portfolio starting from Jan 1990 to Dec 2014.
}
