% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wtsStocks145GmvLo}
\alias{wtsStocks145GmvLo}
\title{Weights data}
\source{
TBA
}
\usage{
data("wtsStocks145GmvLo")
}
\description{
Contains weights obtained after optimizing the portfolio returns of 145 stocks (from dataset stocks145scores6) 
             for a Long-only Min-Var portfolio starting from Jan 2000 to Mar 2013.
}

