% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFfmDT.R
\name{calcFLAM}
\alias{calcFLAM}
\title{calcFLAM}
\usage{
calcFLAM(
  specObj,
  modelStats,
  fitResults,
  analysis = c("ISM", "NEW"),
  targetedVol = 0.06,
  ...
)
}
\arguments{
\item{analysis}{method used in the analysis of fundamental law of active management; one of "none", "ISM", 
or "NEW". Default is "none".}

\item{targetedVol}{numeric; the targeted portfolio volatility in the analysis. Default is 0.06.}
}
\description{
function to calculate fundamental law of active management
}
