% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{factorDataSetDjia}
\alias{factorDataSetDjia}
\title{DJIA stocks Compustat factors 14yrs}
\source{
TBA
}
\usage{
data("factorDataSetDjia")
}
\description{
Contains returns for 30 DJIA stocks spanned across 9 Sectors -ENERGY, COSTAP, INDUS,T MATRLS,
              FINS, INFOTK, HEALTH, CODISC, and TELCOM stocks along with 4 factor data (MKTCAP, ENTVAL, P2B, EV2S, SIZE)
             starting from Jan 2000 to march 2013.
             
             The 9 Sectors correspond to Energy,ConsumerStaples, Industrials, Materials, Financials,
             InformationTechnology, HealthCare, ConsumerDiscretionary and Telecommunications respectively.
}
