% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{factorDataSetDjia5Yrs}
\alias{factorDataSetDjia5Yrs}
\title{DJIA Factor Data}
\source{
TBA
}
\usage{
data("factorDataSetDjia5Yrs")
}
\description{
Contains returns for 30 DJIA stocks along with the 4 factors: MARKETCAP, ENTVALUE, P2B, EV2S, SIZE
             starting from Jan 2008 to Dec 2012.
}

