% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{factorDataSetDjia5Yrs}
\alias{factorDataSetDjia5Yrs}
\title{DJIA stocks Compustat factors 5yrs}
\source{
TBA
}
\usage{
data("factorDataSetDjia5Yrs")
}
\description{
Contains returns for 30 DJIA stocks spanned across 9 Sectors -ENERGY, COSTAP, INDUS,T MATRLS,
              FINS, INFOTK, HEALTH, CODISC, and TELCOM stocks along with 4 factor data (MKTCAP, ENTVAL, P2B, EV2S, SIZE)
              starting from  from Jan 2008 to Dec 2012.
              
              The 9 Sectors correspond to Energy, ConsumerStaples, Industrials, Materials, Financials,
              InformationTechnology, HealthCare, ConsumerDiscretionary and Telecommunications respectively.
}
