% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFfm.R
\name{fitFfm}
\alias{coef.ffm}
\alias{fitFfm}
\alias{fitted.ffm}
\alias{residuals.ffm}
\title{Fit a fundamental factor model using cross-sectional regression}
\usage{
fitFfm(data, asset.var, ret.var, date.var, exposure.vars, weight.var = NULL,
  fit.method = c("LS", "WLS", "Rob", "W-Rob"), rob.stats = FALSE,
  full.resid.cov = FALSE, z.score = FALSE, ...)

\method{coef}{ffm}(object, ...)

\method{fitted}{ffm}(object, ...)

\method{residuals}{ffm}(object, ...)
}
\arguments{
\item{data}{data.frame of the balanced panel data containing the variables 
\code{asset.var}, \code{ret.var}, \code{exposure.vars}, \code{date.var} and 
optionally, \code{weight.var}.}

\item{asset.var}{character; name of the variable for asset names.}

\item{ret.var}{character; name of the variable for asset returns.}

\item{date.var}{character; name of the variable containing the dates 
coercible to class \code{Date}.}

\item{exposure.vars}{vector; names of the variables containing the 
fundamental factor exposures.}

\item{weight.var}{character; name of the variable containing the weights 
used when standarizing style factor exposures. Default is \code{NULL}. See 
Details.}

\item{fit.method}{method for estimating factor returns; one of "LS", "WLS" 
"Rob" or "W-Rob". See details. Default is "LS".}

\item{rob.stats}{logical; If \code{TRUE}, robust estimates of covariance, 
correlation, location and univariate scale are computed as appropriate (see 
Details). Default is \code{FALSE}.}

\item{full.resid.cov}{logical; If \code{TRUE}, a full residual covariance 
matrix is estimated. Otherwise, a diagonal residual covariance matrix is 
estimated. Default is \code{FALSE}.}

\item{z.score}{logical; If \code{TRUE}, style exposures will be converted to 
z-scores; weights given by \code{weight.var}. Default is \code{FALSE}.}

\item{...}{potentially further arguments passed.}

\item{object}{a fit object of class \code{ffm} which is returned by 
\code{fitFfm}}
}
\value{
\code{fitFfm} returns an object of class \code{"ffm"} for which 
\code{print}, \code{plot}, \code{predict} and \code{summary} methods exist. 

The generic accessor functions \code{coef}, \code{fitted} and 
\code{residuals} extract various useful features of the fit object. 
Additionally, \code{fmCov} computes the covariance matrix for asset returns 
based on the fitted factor model.

An object of class \code{"ffm"} is a list containing the following 
components:
\item{factor.fit}{list of fitted objects for each time period. Each object 
is of class \code{lm} if \code{fit.method="LS" or "WLS"}, or, class 
\code{lmRob} if \code{fit.method="Rob" or "W-Rob"}.}
\item{beta}{N x K matrix of factor exposures for the last time period.}
\item{factor.returns}{xts object of K-factor returns (including intercept).}
\item{residuals}{xts object of residuals for N-assets.}
\item{r2}{length-T vector of R-squared values.}
\item{factor.cov}{N x N covariance matrix of the factor returns.}
\item{resid.cov}{N x N covariance matrix of residuals.}
\item{return.cov}{N x N return covariance estimated by the factor model, 
using the factor exposures from the last time period.}
\item{resid.var}{length-N vector of residual variances.}
\item{call}{the matched function call.}
\item{data}{data frame object as input.}
\item{date.var}{date.var as input}
\item{ret.var}{ret.var as input}
\item{asset.var}{asset.var as input.}
\item{exposure.vars}{exposure.vars as input.}
\item{weight.var}{weight.var as input.}
\item{fit.method}{fit.method as input.}
\item{asset.names}{length-N vector of asset names.}
\item{factor.names}{length-K vector of factor.names.}
\item{time.periods}{length-T vector of dates.}
Where N is the number of assets, K is the number of factors (including the 
intercept or dummy variables) and T is the number of unique time periods.
}
\description{
Fit a fundamental (cross-sectional) factor model using ordinary
least squares or robust regression. Fundamental factor models use observable
asset specific characteristics (or) fundamentals, like industry
classification, market capitalization, style classification (value, growth)
etc. to calculate the common risk factors. An object of class \code{"ffm"}
is returned.
}
\details{
Estimation method "LS" corresponds to ordinary least squares using 
\code{\link[stats]{lm}} and "Rob" is robust regression using 
\code{\link[robust]{lmRob}}. "WLS" is weighted least squares using estimates 
of the residual variances from LS regression as weights (feasible GLS). 
Similarly, "W-Rob" is weighted robust regression.

Standardizing style factor exposures: The exposures can be standardized into
z-scores using regular or robust (see \code{rob.stats}) measures of location 
and scale. Further, \code{weight.var}, a variable such as market-cap, can be 
used to compute the weighted mean exposure, and an equal-weighted standard 
deviation of the exposures about the weighted mean. This may help avoid an 
ill-conditioned covariance matrix. Default option equally weights exposures 
of different assets each period. 

If \code{rob.stats=TRUE}, \code{\link[robust]{covRob}} is used to compute a 
robust estimate of the factor covariance/correlation matrix, and, 
\code{\link[robustbase]{scaleTau2}} is used to compute robust tau-estimates 
of univariate scale for residuals during "WLS" or "W-Rob" regressions. When 
standardizing style exposures, the \code{\link[stats]{median}} and 
\code{\link[stats]{mad}} are used for location and scale respectively.

At this time, the regression can contain only one dummy exposure (one of 
industry, sector, country etc.) or intercept term, otherwise the exposure 
matrix will become singular. We plan to expand the function to allow 
specifying more than one dummy variable, and, dummy variable(s) in 
combination with an intercept term in the future. (Ex: Country + Sector + 
Intercept)

The original function was designed by Doug Martin and initially implemented
in S-PLUS by a number of University of Washington Ph.D. students:
Christopher Green, Eric Aldrich, and Yindeng Jiang. Guy Yollin ported the
function to R and Yi-An Chen modified that code. Sangeetha Srinivasan
re-factored, tested, corrected and expanded the functionalities and S3 
methods.
}
\examples{

# Load fundamental and return data
data(Stock.df)

# fit a fundamental factor model
exposure.vars <- c("BOOK2MARKET", "LOG.MARKETCAP")
fit <- fitFfm(data=stock, asset.var="TICKER", ret.var="RETURN", 
              date.var="DATE", exposure.vars=exposure.vars)
names(fit)

# fit a BARRA Industry Factor Model
exposure.vars <- c("GICS.SECTOR")
fit1 <- fitFfm(data=stock, asset.var="TICKER", ret.var="RETURN", 
               date.var="DATE", exposure.vars=exposure.vars, 
               fit.method="Rob", rob.stats=TRUE)

# example with sector dummy included
exposure.vars <- c("BOOK2MARKET", "LOG.MARKETCAP", "GICS.SECTOR")
fit2 <- fitFfm(data=stock, asset.var="TICKER", ret.var="RETURN", 
              date.var="DATE", exposure.vars=exposure.vars)

}
\author{
Sangeetha Srinivasan, Guy Yollin and Yi-An Chen
}
\references{
Menchero, J. (2010). The Characteristics of Factor Portfolios. Journal of
Performance Measurement, 15(1), 52-62.

Grinold, R. C., & Kahn, R. N. (2000). Active portfolio management (Second
Ed.). New York: McGraw-Hill.
}
\seealso{
The \code{ffm} methods for generic functions: 
\code{\link{plot.ffm}}, \code{\link{predict.ffm}}, 
\code{\link{print.ffm}} and \code{\link{summary.ffm}}. 

And, the following extractor functions: \code{\link[stats]{coef}}, 
\code{\link[stats]{fitted}}, \code{\link[stats]{residuals}},
\code{\link{fmCov}}, \code{\link{fmSdDecomp}}, \code{\link{fmVaRDecomp}} 
and \code{\link{fmEsDecomp}}.

\code{\link{paFm}} for Performance Attribution.
}

