% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fitSfm}
\alias{coef.sfm}
\alias{fitSfm}
\alias{fitted.sfm}
\alias{residuals.sfm}
\title{Fit a statistical factor model using principal component analysis}
\usage{
fitSfm(data, k = 1, max.k = NULL, refine = TRUE, sig = 0.05,
  check = FALSE)

\method{coef}{sfm}(object, ...)

\method{fitted}{sfm}(object, ...)

\method{residuals}{sfm}(object, ...)
}
\arguments{
\item{data}{vector, matrix, data.frame, xts, timeSeries or zoo object with
asset returns. See details.}

\item{k}{number of factors; a number (or) a method for determining the
optimal number of factors, one of "bn" or "ck". See details. Default is 1.}

\item{max.k}{scalar; the maximum number of factors to be considered for
methods "bn" or "ck". Default is \code{NULL}. See details.}

\item{refine}{logical; whether to use the Connor-Korajczyk refinement for
APCA. Default is \code{TRUE}.}

\item{sig}{scalar; desired level of significance when "ck" method is
specified. Default is 0.05.}

\item{check}{logical; to check if any asset has identical observations.
Default is \code{FALSE}.}

\item{object}{a fit object of class \code{sfm} which is returned by
\code{fitSfm}}

\item{...}{arguments passed to other functions.}
}
\value{
fitTsfm returns an object of class \code{"sfm"} for which
\code{print}, \code{plot}, \code{predict} and \code{summary} methods exist.

The generic accessor functions \code{coef}, \code{fitted} and
\code{residuals} extract various useful features of the fit object.
Additionally, \code{fmCov} computes the covariance matrix for asset returns
based on the fitted factor model

An object of class \code{"sfm"} is a list containing the following
components:
\item{asset.fit}{list of fitted objects of class \code{lm} for each asset,
from the time-series OLS regression of asset returns on estimated factors.}
\item{k}{number of factors; as input or determined by "ck" or "bn" methods.}
\item{factors}{T x K xts object of estimated factor realizations.}
\item{loadings}{N x K matrix of factor loadings estimated by
regressing the asset returns on estimated factors.}
\item{alpha}{length-N vector of estimated alphas.}
\item{r2}{length-N vector of R-squared values.}
\item{resid.sd}{length-N vector of residual standard deviations.}
\item{residuals}{T x N xts object of residuals from the OLS regression.}
\item{Omega}{N x N return covariance matrix estimated by the factor model.}
\item{eigen}{length-K vector of eigenvalues of the sample covariance matrix.}
\item{mimic}{N x K matrix of factor mimicking portfolio weights.}
\item{call}{the matched function call.}
\item{data}{T x N xts data object containing the asset returns.}
\item{asset.names}{length-N vector of column names from data.}
Where N is the number of assets, K is the number of factors, and T is the
number of observations.

\item{residuals}{T x N matrix of residuals from the regression.}
\item{asset.ret}{N x T matrix of fitted asset returns from the factor model.}
}
\description{
Fits a statistical factor model using principal component
analysis for one or more asset returns or excess returns. When the number of
assets exceeds the number of time periods, APCA (Asymptotic Principal
Component Analysis) is performed. This function is based on the S+FinMetric
function \code{mfactor}. An object of class \code{"sfm"} is returned.
}
\details{
If \code{data} is not of class \code{"xts"}, rownames must provide an
\code{xts} compatible time index. If the data contains missing values,
\code{na.rm} should be set to \code{TRUE} to remove NAs.

Let \code{N} be the number of columns or assets and \code{T} be the number
of rows or observations. When \code{N < T}, Principal Component Analysis
(PCA) is performed. Otherwise, Asymptotic Principal Component Analysis
(APCA) is performed. In either case, any number of factors less than
\code{min(N,T)} can be chosen via argument \code{k}. Default is 1. Refer to
Zivot and Wang (2007) for more details and references.

Alternately, for APCA, a method to determine the number of factors can be
specified: \code{k="bn"} corresponds to Bai and Ng (2002) and \code{k="ck"}
corresponds to Connor and Korajczyk (1993). User can specify the maximum
number of factors, \code{max.k} to consider with these methods. If not, a
default maximum is calculated from \code{min(10, T-1)}.

\code{refine} specifies whether a refinement of the APCA procedure (that may
improve efficiency) from Connor and Korajczyk (1988) is to be used.

If \code{check=TRUE}, a warning is issued if any asset is found to have
identical observations.

Note about NAs: Before model fitting, incomplete cases in \code{data} are
removed using \code{\link[stats]{na.omit}}. Otherwise, all observations are
included.
}
\examples{
# load data for fitSfm.r
data(stat.fm.data)
# data is from finmetric berndt.dat and folio.dat

# PCA is performed on sfm.dat and APCA on sfm.apca.dat
class(sfm.dat)
class(sfm.apca.dat)

# pca
args(fitSfm)
sfm.pca.fit <- fitSfm(sfm.dat, k=2)
class(sfm.pca.fit)
names(sfm.pca.fit)
head(sfm.pca.fit$factors)
head(sfm.pca.fit$loadings)
sfm.pca.fit$r2
sfm.pca.fit$resid.sd
sfm.pca.fit$mimic

# apca with number of factors, k=15
sfm.apca.fit <- fitSfm(sfm.apca.dat, k=15, refine=TRUE)

# apca with the Bai & Ng method
sfm.apca.fit.bn <- fitSfm(sfm.apca.dat, k="bn")

# apca with the Connor-Korajczyk method
sfm.apca.fit.ck <- fitSfm(sfm.apca.dat, k="ck")
}
\author{
Eric Zivot, Sangeetha Srinivasan and Yi-An Chen
}
\references{
Bai, J., & Ng, S. (2002). Determining the number of factors in approximate
factor models. Econometrica, 70(1), 191-221.

Connor, G., & Korajczyk, R. A. (1988). Risk and return in an equilibrium
APT: Application of a new test methodology. Journal of Financial Economics,
21(2), 255-289.

Connor, G., & Korajczyk, R. A. (1993). A test for the number of factors in
an approximate factor model. The Journal of Finance, 48(4), 1263-1291.

Zivot, E., & Wang, J. (2007). Modeling Financial Time Series with S-PLUS
(Vol. 191). Springer.
}
\seealso{
The \code{sfm} methods for generic functions:
\code{\link{plot.sfm}}, \code{\link{predict.sfm}},
\code{\link{print.sfm}} and \code{\link{summary.sfm}}.

And, the following extractor functions: \code{\link[stats]{coef}},
\code{\link[stats]{fitted}}, \code{\link[stats]{residuals}},
\code{\link{fmCov}}, \code{\link{fmSdDecomp}}, \code{\link{fmVaRDecomp}}
and \code{\link{fmEsDecomp}}.

\code{\link{paFm}} for Performance Attribution.
}

