% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitStatisticalFactorModel}
\alias{fitStatisticalFactorModel}
\title{Fit a statistical factor model using principal component analysis}
\usage{
fitStatisticalFactorModel(data, k = 1, refine = TRUE, check = FALSE,
  max.k = NULL, sig = 0.05, na.rm = FALSE)
}
\arguments{
\item{data}{a vector, matrix, data.frame, xts, timeSeries or zoo object with
asset returns and factors names. If data is not of class xts, rownames must
provide an xts compatible time index.}

\item{k}{numbers of factors. Can be a scalar value or a method for
determining the optimal number of factors. k="bn" corresponds to Bai and
Ng (2002) and k="ck" corresponds to Connor and Korajczyk (1993). Defaults to 1.}

\item{refine}{a logical value that when set to \code{TRUE}, specifies the
Connor-Korajczyk refinement for APCA (Asymptotic Principal Component Analysis).
Defaults to \code{TRUE}.}

\item{check}{Checks if any two assets have identical values. Defaults to
\code{FALSE}.}

\item{max.k}{a scalar that specifies the maximum number of factors to be
considered.}

\item{sig}{desired level of significant when "ck"" method is specified.}

\item{na.rm}{a logical value to specify if missing values should be removed.
Defaults to FALSE.}
}
\value{
\itemize{
\item{factors}{ T x K matrix of estimated factors.}
\item{loadings}{ K x N matrix of asset specific factor loadings beta_i,
estimated by regressing the asset returns on factors.}
\item{alpha}{ 1 x N vector of estimated intercepts alpha_i}
\item{ret.cov}{ N x N matrix of asset returns' sample covariance matrix.}
\item{r2}{ r-squared value from regressing the asset returns on the factors.}
\item{k}{ the number of facotrs.}
\item{eigen}{ eigenvalues from the sample covariance matrix.}
\item{residuals}{ T x N matrix of residuals from regression.}
\item{asset.ret}{ asset returns}
\item{asset.fit}{ List of regression lm class of individual returns on
factors.}
\item{resid.variance}{ vector of residual variances.}
\item{mimic}{ N x K matrix of factor mimicking portfolio returns.}
}
Where N is the number of assets, K is the number of factors, and T is the
number of observations.
}
\description{
Fits a statistical factor model using principal component analysis.
This is an adaptation of the S+FinMetric function \code{mfactor}.
}
\examples{
# load data for fitStatisticalFactorModel.r
# data from finmetric berndt.dat and folio.dat

data(stat.fm.data)
##
# sfm.dat is for pca
# sfm.apca.dat is for apca
class(sfm.dat)
class(sfm.apca.dat)

# pca
args(fitStatisticalFactorModel)
sfm.pca.fit <- fitStatisticalFactorModel(sfm.dat,k=2)
class(sfm.pca.fit)
names(sfm.pca.fit)
sfm.pca.fit$factors
sfm.pca.fit$loadings
sfm.pca.fit$r2
sfm.pca.fit$residuals
sfm.pca.fit$resid.variance
sfm.pca.fit$mimic
# apca
sfm.apca.fit <- fitStatisticalFactorModel(sfm.apca.dat,k=1)
names(sfm.apca.fit)
sfm.apca.res <- sfm.apca.fit$residuals
sfm.apca.mimic <- sfm.apca.fit$mimic
# apca with bai and Ng method
sfm.apca.fit.bn <- fitStatisticalFactorModel(sfm.apca.dat,k="bn")
class(sfm.apca.fit.bn)
names(sfm.apca.fit.bn)
sfm.apca.fit.bn$mimic

# apca with ck method
sfm.apca.fit.ck <- fitStatisticalFactorModel(sfm.apca.dat,k="ck")
class(sfm.apca.fit.ck)
names(sfm.apca.fit.ck)
sfm.apca.fit.ck$mimic
}
\author{
Eric Zivot and Yi-An Chen
}
\references{
Zivot and Wang, (2006) "Modeling Financial Time Series with S-PLUS, 2nd edition"
}

