% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitTSFM}
\alias{fitTSFM}
\title{Fit a time series factor model using time series regression}
\usage{
fitTSFM(asset.names, factor.names, market.name, data = data,
  fit.method = c("OLS", "DLS", "Robust"), variable.selection = c("none",
  "stepwise", "all subsets", "lars", "lasso"),
  subsets.method = c("exhaustive", "backward", "forward", "seqrep"),
  nvmax = 8, force.in = NULL, num.factors.subset = 1,
  add.up.market = FALSE, add.market.sqd = FALSE, decay = 0.95,
  lars.criterion = "Cp", ...)
}
\arguments{
\item{asset.names}{vector containing names of assets, whose returns or
excess returns are the dependent variable.}

\item{factor.names}{vector containing names of the macroeconomic factors.}

\item{market.name}{name of the column for market excess returns (Rm-Rf).
Is required only if \code{add.up.market} or \code{add.up.market.squared}
are \code{TRUE}.}

\item{data}{vector, matrix, data.frame, xts, timeSeries or zoo object
containing column(s) named \code{asset.names}, \code{factor.names} and
optionally, \code{market.name}.}

\item{fit.method}{the estimation method, one of "OLS", "DLS" or "Robust".
See details.}

\item{variable.selection}{the variable selection method, one of "none",
"stepwise","all subsets","lars" or "lasso". See details.}

\item{subsets.method}{one of "exhaustive", "forward", "backward" or "seqrep"
(sequential replacement) to specify the type of subset search/selection.
Required if "all subsets" variable selection is chosen.}

\item{nvmax}{the maximum size of subsets to examine; an option for
"all subsets" variable selection. Default is 8.}

\item{force.in}{vector containing the names of factors that should always
be included in the model; an option for "all subsets" variable selection.
Default is NULL.}

\item{num.factors.subset}{number of factors required in the factor model;
an option for "all subsets" variable selection. Default is 1.
Note: nvmax >= num.factors.subset >= length(force.in).}

\item{add.up.market}{logical; If \code{TRUE}, adds max(0, Rm-Rf) as a
regressor and \code{market.name} is also required. Default is \code{FALSE}.
See Details.}

\item{add.market.sqd}{logical; If \code{TRUE}, adds (Rm-Rf)^2 as a
regressor and \code{market.name} is also required. Default is \code{FALSE}.}

\item{decay}{a scalar in (0, 1] to specify the decay factor for
\code{fit.method="DLS"}. Default is 0.95.}

\item{lars.criterion}{an option to assess model selection for the "lars" or
"lasso" variable.selection methods; one of "Cp" or "cv". See details.
Default is "Cp".}

\item{...}{optional arguments passed to the \code{step} function for
variable.selection method "stepwise", such as direction, steps and
the penalty factor k. Note that argument k is available only for "OLS"
and "DLS" fits. Scope argument is not available presently. Also plan to
include other controls passed to \code{lmRob} soon.}
}
\value{
fitTSFM returns an object of class
\code{tsfm}. The returned object is a list
containing the following components:
\item{asset.fit}{list of fitted objects for each asset. Each object is of
class \code{lm} if \code{fit.method="OLS" or "DLS"}, class \code{lmRob} if
the \code{fit.method="Robust"}, or class \code{lars} if
\code{variable.selection="lars" or "lasso"}.}
\item{alpha}{N x 1 vector of estimated alphas.}
\item{beta}{N x K matrix of estimated betas.}
\item{r2}{N x 1 vector of R-squared values.}
\item{resid.sd}{N x 1 vector of residual standard deviations.}
\item{call}{the matched function call.}
\item{data}{data as input.}
\item{asset.names}{asset.names as input.}
\item{factor.names}{factor.names as input.}
\item{fit.method}{fit.method as input.}
\item{variable.selection}{variable.selection as input.}
Where N is the number of assets and K is the number of factors.
}
\description{
Fits a time series (or, macroeconomic) factor model for single
or multiple asset returns or excess returns using time series regression.
Users can choose between ordinary least squares-OLS, discounted least
squares-DLS (or) robust regression. Several variable selection options
including Stepwise, Subsets, Lars are available as well. An object of class
\code{tsfm} is returned.
}
\details{
Estimation method "OLS" corresponds to ordinary least squares, "DLS" is
discounted least squares, which is weighted least squares estimation with
exponentially declining weights that sum to unity, and, "Robust" is robust
regression (uses \code{\link[robust]{lmRob}}).

If \code{variable.selection="none"}, all chosen factors are used in the
factor model. Whereas, "stepwise" performs traditional forward/backward
stepwise OLS regression (using \code{\link[stats]{step}}), that starts from
the initial set of factors and adds factors only if the regression fit, as
measured by the Bayesian Information Criterion (BIC) or Akaike Information
Criterion (AIC), improves. And, "all subsets" enables subsets selection
using \code{\link[leaps]{regsubsets}} that chooses the n-best performing
subsets of any given size (specified as \code{num.factor.subsets} here).
"lars" and "lasso" correspond to variants of least angle regression using
\code{\link[lars]{lars}}. If "lars" or "lasso" are chosen, \code{fit.method}
will be ignored.

Note: If  \code{variable.selection}="lars" or "lasso", \code{fit.method}
will be ignored. And, "Robust" \code{fit.method} is not truly available with
\code{variable.selection="all subsets"}; instead, results are produced for
\code{variable.selection="none"} with "Robust" to include all factors.

If \code{add.up.market = TRUE}, max(0, Rm-Rf) is added as a factor in the
regression, following Henriksson & Merton (1981), to account for market
timing (price movement of the general stock market relative to fixed income
securities). The coefficient can be interpreted as the number of free put
options.

Finally, for both the "lars" and "lasso" methods, the "Cp" statistic
(defined in page 17 of Efron et al. (2002)) is calculated using
\code{\link[lars]{summary.lars}} . While, "cv" computes the K-fold
cross-validated mean squared prediction error using
\code{\link[lars]{cv.lars}}.
}
\examples{
# load data from the database
data(managers.df)
fit <- fitTSFM(asset.names=colnames(managers.df[,(1:6)]),
               factor.names=c("EDHEC.LS.EQ","SP500.TR"), data=managers.df,
               fit.method="OLS", variable.selection="none")
# summary of HAM1
summary(fit$asset.fit$HAM1)
# plot actual vs. fitted over time for HAM1
# using chart.TimeSeries() function from PerformanceAnalytics package
dataToPlot <- cbind(fitted(fit$asset.fit$HAM1), na.omit(managers.df$HAM1))
colnames(dataToPlot) <- c("Fitted","Actual")
chart.TimeSeries(dataToPlot, main="FM fit for HAM1",
                 colorset=c("black","blue"), legend.loc="bottomleft")
}
\author{
Eric Zivot, Yi-An Chen and Sangeetha Srinivasan.
}
\references{
\enumerate{
\item Christopherson, Carino and Ferson (2009). Portfolio Performance
Measurement and Benchmarking, McGraw Hill.
\item Efron, Hastie, Johnstone and Tibshirani (2002) "Least Angle
Regression" (with discussion) Annals of Statistics. Also refer to
\url{http://www-stat.stanford.edu/~hastie/Papers/LARS/LeastAngle_2002.pdf}.
\item Hastie, Tibshirani and Friedman (2008) Elements of Statistical
Learning 2nd edition, Springer, NY.
\item Henriksson and Merton (1981). On market timing and investment
performance. II. Statistical procedures for evaluating forecasting skills,
Journal of Business, Vol 54, No 4.
}
}
\seealso{
The following generic method functions: \code{\link{plot.tsfm}},
\code{\link{predict.tsfm}}, \code{\link{print.tsfm}} and
\code{\link{summary.tsfm}}.

And, the following extractor functions: \code{\link{coef.tsfm}},
\code{\link{cov.tsfm}}, \code{\link{fitted.tsfm}} and
\code{\link{residuals.tsfm}}.

\code{\link{paFM}} for Performance Attribution.
}

