% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitTsfm}
\alias{coef.tsfm}
\alias{covFm.tsfm}
\alias{fitTsfm}
\alias{fitted.tsfm}
\alias{residuals.tsfm}
\title{Fit a time series factor model using time series regression}
\usage{
fitTsfm(asset.names, factor.names, mkt.name = NULL, rf.name = NULL,
  data = data, fit.method = c("OLS", "DLS", "Robust"),
  variable.selection = c("none", "stepwise", "subsets", "lars"),
  mkt.timing = NULL, control = fitTsfm.control(...), ...)

\method{coef}{tsfm}(object, ...)

\method{fitted}{tsfm}(object, ...)

\method{residuals}{tsfm}(object, ...)

\method{covFm}{tsfm}(object)
}
\arguments{
\item{asset.names}{vector containing names of assets, whose returns or
excess returns are the dependent variable.}

\item{factor.names}{vector containing names of the macroeconomic factors.}

\item{mkt.name}{name of the column for market excess returns (Rm-Rf).
Is required only if \code{add.up.market} or \code{add.market.sqd}
are \code{TRUE}. Default is NULL.}

\item{rf.name}{name of the column of risk free rate variable to calculate
excess returns for all assets (in \code{asset.names}) and factors (in
\code{factor.names}). Default is NULL, and no action is taken.}

\item{data}{vector, matrix, data.frame, xts, timeSeries or zoo object
containing column(s) named in \code{asset.names}, \code{factor.names} and
optionally, \code{mkt.name} and \code{rf.name}.}

\item{fit.method}{the estimation method, one of "OLS", "DLS" or "Robust".
See details. Default is "OLS".}

\item{variable.selection}{the variable selection method, one of "none",
"stepwise","subsets","lars". See details. Default is "none".}

\item{mkt.timing}{one of "HM", "TM" or "both". Default is NULL. See Details.
\code{mkt.name} is required if any of these options are specified.}

\item{control}{list of control parameters. The default is constructed by
the function \code{\link{fitTsfm.control}}. See the documentation for
\code{\link{fitTsfm.control}} for details.}

\item{...}{arguments passed to \code{\link{fitTsfm.control}}}

\item{object}{a fit object of class \code{tsfm} which is returned by
\code{fitTsfm}}
}
\value{
fitTsfm returns an object of class \code{tsfm}.

The generic functions \code{summary}, \code{predict} and \code{plot} are
used to obtain and print a summary, predicted asset returns for new factor
data and plot selected characteristics for one or more assets. The generic
accessor functions \code{coef}, \code{fitted} and \code{residuals}
extract various useful features of the fit object. \code{coef.tsfm} extracts
coefficients from the fitted factor model and returns an N x (K+1) matrix of
all coefficients, \code{fitted.tsfm} gives an N x T data object of fitted
values and \code{residuals.tsfm} gives an N x T data object of residuals.
Additionally, \code{covFm} computes the \code{N x N} covariance matrix for
asset returns based on the fitted factor model

An object of class \code{tsfm} is a list containing the following
components:
\item{asset.fit}{list of fitted objects for each asset. Each object is of
class \code{lm} if \code{fit.method="OLS" or "DLS"}, class \code{lmRob} if
the \code{fit.method="Robust"}, or class \code{lars} if
\code{variable.selection="lars"}.}
\item{alpha}{N x 1 vector of estimated alphas.}
\item{beta}{N x K matrix of estimated betas.}
\item{r2}{N x 1 vector of R-squared values.}
\item{resid.sd}{N x 1 vector of residual standard deviations.}
\item{fitted}{xts data object of fitted values; iff
\code{variable.selection="lars"}}
\item{call}{the matched function call.}
\item{data}{xts data object containing the assets and factors.}
\item{asset.names}{asset.names as input.}
\item{factor.names}{factor.names as input.}
\item{fit.method}{fit.method as input.}
\item{variable.selection}{variable.selection as input.}
Where N is the number of assets, K is the number of factors and T is the
number of time periods.
}
\description{
Fits a time series (or, macroeconomic) factor model for one
or more asset returns (or, excess returns) using time series regression.
Users can choose between ordinary least squares-OLS, discounted least
squares-DLS (or) robust regression. Several variable selection options
including Stepwise, Subsets, Lars are available as well. An object of class
\code{tsfm} is returned.
}
\details{
Typically, factor models are fit using excess returns. \code{rf.name} gives
the option to supply a risk free rate variable to subtract from each asset
return and factor to create excess returns.

Estimation method "OLS" corresponds to ordinary least squares using
\code{\link[stats]{lm}}, "DLS" is discounted least squares (weighted least
squares with exponentially declining weights that sum to unity), and,
"Robust" is robust regression (useing \code{\link[robust]{lmRob}}).

If \code{variable.selection="none"}, all chosen factors are used in the
factor model. Whereas, "stepwise" performs traditional forward/backward
stepwise OLS regression (using \code{\link[stats]{step}} or
\code{\link[robust]{step.lmRob}}), that starts from the initial set of
factors and adds factors only if the regression fit, as measured by the
Bayesian Information Criterion (BIC) or Akaike Information Criterion (AIC),
improves. And, "subsets" enables subsets selection using
\code{\link[leaps]{regsubsets}}; chooses the best performing subset of any
given size. See \code{\link{fitTsfm.control}} for more details on the
control arguments. \code{varaible.selection="lars"} corresponds to least
angle regression using \code{\link[lars]{lars}} with variants "lasso",
"lar", "forward.stagewise" or "stepwise". Note: If
\code{variable.selection="lars"}, \code{fit.method} will be ignored.

\code{mkt.timing} allows for market-timing factors to be added to any of the
above methods. Market timing accounts for the price movement of the general
stock market relative to fixed income securities). "HM" follows
Henriksson & Merton (1981) and \code{up-market = max(0, Rm-Rf)}, is added
as a factor in the regression. The coefficient of this up-market factor can
be interpreted as the number of free put options. Similarly, "TM" follows
Treynor-Mazuy (1966), to account for market timing with respect to
volatility, and \code{market.sqd = (Rm-Rf)^2} is added as a factor in the
regression. Option "both" adds both of these factors.

Note about NAs: Before model fitting, incomplete cases are removed for
every asset (return data combined with respective factors' return data)
using \code{\link[stats]{na.omit}}. Otherwise, all observations in
\code{data} are included.

Note about spaces in asset/factor names: Spaces in column names of the data
object will be converetd to periods as the function works with \code{xts}
objects internally and hence column names can't be retained as such.
}
\examples{
# load data from the database
data(managers)
fit <- fitTsfm(asset.names=colnames(managers[,(1:6)]),
               factor.names=colnames(managers[,(7:9)]),
               mkt.name="SP500 TR", mkt.timing="both", data=managers)
# summary
summary(fit)
# fitted values for all assets' returns
fitted(fit)
# plot actual vs. fitted over time for HAM1
# using chart.TimeSeries() function from PerformanceAnalytics package
dataToPlot <- cbind(fitted(fit$asset.fit$HAM1), na.omit(managers$HAM1))
colnames(dataToPlot) <- c("Fitted","Actual")
chart.TimeSeries(dataToPlot, main="FM fit for HAM1",
                 colorset=c("black","blue"), legend.loc="bottomleft")
}
\author{
Eric Zivot, Yi-An Chen and Sangeetha Srinivasan.
}
\references{
\enumerate{
\item Christopherson, Jon A., David R. Carino, and Wayne E. Ferson.
Portfolio performance measurement and benchmarking. McGraw Hill
Professional, 2009.
\item Efron, Bradley, Trevor Hastie, Iain Johnstone, and Robert Tibshirani.
"Least angle regression." The Annals of statistics 32, no.2 (2004): 407-499.
\item Hastie, Trevor, Robert Tibshirani, Jerome Friedman, T. Hastie, J.
Friedman, and R. Tibshirani. The elements of statistical learning. Vol. 2,
no. 1. New York: Springer, 2009.
\item Henriksson, Roy D., and Robert C. Merton. "On market timing and
investment performance. II. Statistical procedures for evaluating
forecasting skills." Journal of business (1981): 513-533.
\item Treynor, Jack, and Kay Mazuy. "Can mutual funds outguess the market."
Harvard business review 44, no. 4 (1966): 131-136.
}
}
\seealso{
The \code{tsfm} methods for generic functions:
\code{\link{plot.tsfm}}, \code{\link{predict.tsfm}},
\code{\link{print.tsfm}} and \code{\link{summary.tsfm}}.

And, the following extractor functions: \code{\link[stats]{coef}},
\code{\link{covFm}}, \code{\link[stats]{fitted}} and
\code{\link[stats]{residuals}}.

\code{\link{paFm}} for Performance Attribution.
}

