\name{managers.df}
\alias{managers.df}
\docType{data}
\title{
Hypothetical Alternative Asset Manager and Benchmark Data
}
\description{
a data.frame format from managers dataset from package PerformanceAnalytics, containing columns of monthly returns for six hypothetical asset managers (HAM1 through HAM6), the EDHEC Long-Short Equity hedge fund index, the S\&P 500 total returns. Monthly returns for all series end in December 2006 and begin at different periods starting from January 1997.
}
\usage{data(managers.df)}

\examples{
data(managers.df)
## maybe str(managers.df) ; plot(managers.df) ...
}
\keyword{datasets}
