% Generated by roxygen2 (4.0.1): do not edit by hand
\name{paFM}
\alias{paFM}
\title{Compute performance attribution in factor models}
\usage{
paFM(fit, ...)
}
\arguments{
\item{fit}{an object of class \code{tsfm}, \code{sfm} or \code{ffm}.}

\item{...}{other arguments/controls passed to the fit methods.}
}
\value{
The returned object is of class \code{"pafm"} containing
\describe{
\item{cum.ret.attr.f}{N X J matrix of cumulative return attributed to
factors.}
\item{cum.spec.ret}{1 x N vector of cumulative specific returns.}
\item{attr.list}{list of time series of attributed returns for every
portfolio.}
}
}
\description{
Decompose total returns into returns attributed to factors and
specific returns. An object of class \code{"pafm"} is generated and generic
functions such as \code{plot}, \code{summary} and \code{print} can be used.
}
\details{
Total returns can be decomposed into returns attributed to factors
and specific returns. \cr \eqn{R_t = \sum  b_j * f_jt + u_t,t=1...T} \cr
\code{b_j} is exposure to factor j and \code{f_jt} is factor j.
The returns attributed to factor j is \code{b_j * f_jt} and specific
returns is \code{u_t}.
}
\examples{
\dontrun{
data(managers.df)
fit.ts <- fitTSFM(assets.names=colnames(managers.df[, (1:6)]),
                  factors.names=c("EDHEC.LS.EQ","SP500.TR"),
                  data=managers.df, fit.method="OLS")
# without benchmark
fm.attr <- paFM(fit.ts)
}
}
\author{
Yi-An Chen and Sangeetha Srinivasan
}
\references{
Grinold, R. and Kahn, R. \emph{Active Portfolio Management},
McGraw-Hill.
}
\seealso{
\code{\link{fitTSFM}}, \code{\link{fitSFM}}, \code{\link{fitFFM}}
}

