% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.tsfm}
\alias{plot.tsfm}
\title{Plots from a fitted time series factor model}
\usage{
\method{plot}{tsfm}(x, which = NULL, max.show = 6, plot.single = FALSE,
  asset.name, colorset = (1:12), legend.loc = "topleft", las = 1,
  VaR.method = "historical", ...)
}
\arguments{
\item{x}{an object of class \code{tsfm} produced by \code{fitTsfm}.}

\item{which}{a number to indicate the type of plot. If a subset of the plots
is required, specify a subset of the numbers 1:10 for group plots and 1:16
for individual plots. If \code{which=NULL} (default), the following menu
appears: \cr \cr
For plots of a group of assets: \cr
1 = Factor model coefficients: Alpha, \cr
2 = Factor model coefficients: Betas, \cr
3 = Actual and Fitted asset returns, \cr
4 = R-squared, \cr
5 = Residual Volatility,\cr
6 = Factor Model Residual Correlation \cr
7 = Factor Model Correlation,\cr
8 = Factor Contribution to SD,\cr
9 = Factor Contribution to ES,\cr
10 = Factor Contribution to VaR \cr \cr
For individual asset plots:\cr
1 = Actual and fitted returns,\cr
2 = Residuals and fitted returns, \cr
3 = Scale-Location plot, \cr
4 = Residuals with standard error bands, \cr
5 = Time series of squared residuals, \cr
6 = Time series of absolute residuals,\cr
7 = SACF and PACF of residuals,\cr
8 = SACF and PACF of squared residuals,\cr
9 = SACF and PACF of absolute residuals,\cr
10 = Density estimate of residuals, \cr
11 = Histogram of residuals with normal curve overlayed,\cr
12 = Normal QQ-plot of residuals,\cr
13 = CUSUM test-Recursive residuals,\cr
14 = CUSUM test-OLS residuals,\cr
15 = Recursive estimates (RE) test of OLS regression coefficients,\cr
16 = Rolling estimates over a 24-period observation window}

\item{max.show}{maximum number of assets in a given plot. Default is 6.}

\item{plot.single}{logical; If \code{TRUE} plots the characteristics of an
individual asset's factor model. The type of plot is given by
\code{which}. Default is \code{FALSE}.}

\item{asset.name}{name of the individual asset to be plotted. Is necessary
if \code{x} contains multiple asset fits and \code{plot.single=TRUE}.}

\item{colorset}{color palette to use for all the plots. Default is
\code{c(1:12)}. The 1st element will be used for individual time series
plots or the 1st series plotted, the 2nd element for the 2nd object in the
plot and so on.}

\item{legend.loc}{places a legend into one of nine locations on the chart:
"bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright",
"right", or "center". Default is "bottomright". Use \code{legend.loc=NULL}
to suppress the legend.}

\item{las}{one of {0, 1, 2, 3} to set the direction of axis labels, same as
in \code{plot}. Default here is 1.}

\item{VaR.method}{a method for computing VaR; one of "modified", "gaussian",
"historical" or "kernel". VaR is computed using
\code{\link[PerformanceAnalytics]{VaR}}. Default is "historical".}

\item{...}{further arguments to be passed to other plotting functions.}
}
\description{
Generic \code{plot} method for object of class \code{tsfm}.
Plots chosen characteristic(s) for one or more assets.
}
\details{
The function can be used for group plots and individual plots. User can
select the type of plot either from the menu prompt (default) or directly
via argument \code{which}.

In case multiple plots are needed, the menu is repeated after each plot
(enter 0 to exit). User can also input a numeric vector of plot options via
\code{which}.

For group plots (the default), the first \code{max.show} assets are plotted.

Setting \code{plot.single=TRUE} enables individual plots. If there is more
than one asset fit by \code{x}, \code{asset.name} should be specified.
However, if the \code{tsfm} object \code{x} only contains one asset’s factor
model fit, plot.tsfm can infer \code{asset.name} without user input.

CUSUM plots (individual asset plot options 10, 11 and 12) are applicable
only for \code{fit.method="OLS"}.

Rolling estimates (individual asset plot option 13) is not applicable for
\code{variable.slection="lars"}.
}
\examples{
# load data from the database
data(managers)
fit.macro <- fitTsfm(asset.names=colnames(managers[,(1:6)]),
                     factor.names=colnames(managers[,(7:8)]),
                     rf.name="US.3m.TR", data=managers)

# for group plots (default), user can select type from menu prompt
# menu is repeated to get multiple types of plots based on the same fit
# plot(fit.macro)

# plot specific option(s) though which
# plot the factor betas of 1st 4 assets fitted above
plot(fit.macro, max.show=4, which=2)

# plot factor model return correlation; angular order of the eigenvectors
plot(fit.macro, which=7, order="AOE", method="ellipse", tl.pos = "d")

# for individual plots: set plot.single=TRUE; specify asset.name
# histogram of residuals from an individual asset's factor model fit
plot(fit.macro, plot.single=TRUE, asset.name="HAM1", which=8)
}
\author{
Eric Zivot, Sangeetha Srinivasan and Yi-An Chen
}
\seealso{
\code{\link{fitTsfm}} and \code{\link{summary.tsfm}} for details
about the time series factor model fit, extractor functions and summary
statistics.

\code{\link[strucchange]{efp}} for CUSUM tests.

\code{\link[xts]{plot.xts}},
\code{\link[PerformanceAnalytics]{chart.TimeSeries}},
\code{\link[PerformanceAnalytics]{chart.ACFplus}},
\code{\link[PerformanceAnalytics]{chart.Histogram}},
\code{\link[PerformanceAnalytics]{chart.QQPlot}},
\code{\link[graphics]{barplot}}, \code{\link[lattice]{barchart}} and
\code{\link[corrplot]{corrplot}} for plotting methods used.

\code{\link{fmSdDecomp}}, \code{\link{fmEsDecomp}},
\code{\link{fmVaRDecomp}} for factor model risk measures.
}

