% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.tsfm}
\alias{plot.tsfm}
\title{Plots from a fitted time series factor model}
\usage{
\method{plot}{tsfm}(x, colorset = c(1:12), legend.loc = NULL,
  which.plot = c("none", "1L", "2L", "3L", "4L", "5L", "6L", "7L"),
  max.show = 6, plot.single = FALSE, asset.name,
  which.plot.single = c("none", "1L", "2L", "3L", "4L", "5L", "6L", "7L",
  "8L", "9L", "10L", "11L", "12L", "13L"), VaR.method = "historical", ...)
}
\arguments{
\item{x}{an object of class \code{tsfm} produced by \code{fitTSFM}.}

\item{colorset}{a vector of colors for the bars or bar components. Argument
is used by \code{\link[graphics]{barplot}}. Default is c(1:12).}

\item{legend.loc}{places a legend into one of nine locations on the chart:
bottomright, bottom, bottomleft, left, topleft, top, topright, right, or
center. Argument is used by
\code{\link[PerformanceAnalytics]{chart.TimeSeries}}. Default is \code{NULL}.}

\item{which.plot}{a number or "none" to indicate which type of group plot to
create for multiple assets. Default is "none"; which brings up the following
menu to select a type. \cr
1 = "Fitted asset returns", \cr
2 = "R-squared", \cr
3 = "Residual Volatility",\cr
4 = "FM Correlation",\cr
5 = "Factors' Contribution to SD",\cr
6 = "Factors' Contribution to ES",\cr
7 = "Factors' Contribution to VaR"}

\item{max.show}{maximum number of assets in a plot. Default is 6.}

\item{plot.single}{a logical value. If \code{TRUE}, plots an individual
asset's linear factor model trait selected by \code{which.plot.single}.
Default is \code{FALSE}.}

\item{asset.name}{name of the individual asset to be plotted. Is necessary
if \code{plot.single=TRUE}}

\item{which.plot.single}{a number or "none" to indicate which type of group
plot to create for multiple assets. Default is "none"; which brings up the
following menu to select a type.\cr
 1 = time series plot of actual and fitted factor returns,\cr
 2 = time series plot of residuals with standard error bands, \cr
 3 = time series plot of squared residuals, \cr
 4 = time series plot of absolute residuals,\cr
 5 = SACF and PACF of residuals,\cr
 6 = SACF and PACF of squared residuals,\cr
 7 = SACF and PACF of absolute residuals,\cr
 8 = histogram of residuals with normal curve overlayed,\cr
 9 = normal qq-plot of residuals,\cr
 10= CUSUM plot of recursive residuals,\cr
 11= CUSUM plot of OLS residuals,\cr
 12= CUSUM plot of recursive estimates relative to full sample estimates,\cr
 13= rolling estimates over a 24-period observation window}

\item{VaR.method}{a method for computing VaR; one of "modified", "gaussian",
"historical" or "kernel". VaR is computed using
\code{\link[PerformanceAnalytics]{VaR}}. Default is "historical".}

\item{...}{further arguments passed to or from other methods.}
}
\description{
S3 \code{plot} method for object of class \code{tsfm}. Plots
selected characteristics for one or more assets.
}
\examples{
\dontrun{
# load data from the database
data(managers.df)
fit.macro <- fitTSFM(asset.names=colnames(managers.df[,(1:6)]),
                     factor.names=c("EDHEC.LS.EQ","SP500.TR"),
                     data=managers.df, fit.method="OLS",
                     variable.selection="none")
# plot all assets and show only the first 4 assets.
plot(fit.macro,max.show=4)
# plot of an individual asset, "HAM1"
plot(fit.macro, plot.single=TRUE, asset.name="HAM1")
}
}
\author{
Eric Zivot, Yi-An Chen and Sangeetha Srinivasan
}
\seealso{
\code{\link{fitTSFM}}, \code{\link{summary.tsfm}}
}

