% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.tsfm.r
\name{plot.tsfm}
\alias{plot.tsfm}
\title{Plots from a fitted time series factor model}
\usage{
\method{plot}{tsfm}(x, which.plot.group = NULL, max.show = 6,
  plot.single = FALSE, asset.name, which.plot.single = NULL,
  colorset = (1:12), legend.loc = "topleft", las = 1,
  VaR.method = "historical", loop = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{tsfm} produced by \code{fitTsfm}.}

\item{which.plot.group}{a number to indicate the type of group plot for
multiple assets. If \code{NULL} (default), the following menu appears: \cr
1 = Factor model coefficients: Alpha, \cr
2 = Factor model coefficients: Betas, \cr
3 = Actual and Fitted asset returns, \cr
4 = R-squared, \cr
5 = Residual Volatility,\cr
6 = Factor Model Residual Correlation \cr
7 = Factor Model Correlation,\cr
8 = Factor Contribution to SD,\cr
9 = Factor Contribution to ES,\cr
10 = Factor Contribution to VaR}

\item{max.show}{maximum number of assets in a given plot. Default is 6.}

\item{plot.single}{logical; If \code{TRUE} plots the characteristics of an
individual asset's factor model. The type of plot is given by
\code{which.plot.single}. Default is \code{FALSE}.}

\item{asset.name}{name of the individual asset to be plotted. Is necessary
if \code{x} contains multiple asset fits and \code{plot.single=TRUE}.}

\item{which.plot.single}{a number to indicate the type of group plot for an
individual asset. If \code{NULL} (default), the following menu appears: \cr
 1 = Time series plot of actual and fitted asset returns,\cr
 2 = Time series plot of residuals with standard error bands, \cr
 3 = Time series plot of squared residuals, \cr
 4 = Time series plot of absolute residuals,\cr
 5 = SACF and PACF of residuals,\cr
 6 = SACF and PACF of squared residuals,\cr
 7 = SACF and PACF of absolute residuals,\cr
 8 = Histogram of residuals with normal curve overlayed,\cr
 9 = Normal qq-plot of residuals,\cr
 10 = CUSUM test-Recursive residuals,\cr
 11 = CUSUM test-OLS residuals,\cr
 12 = Recursive estimates (RE) test of OLS regression coefficients,\cr
 13 = Rolling estimates over a 24-period observation window}

\item{colorset}{color palette to use for all the plots. Default is
\code{c(1:12)}. The 1st element will be used for individual time series
plots or the 1st series plotted, the 2nd element for the 2nd object in the
plot and so on.}

\item{legend.loc}{places a legend into one of nine locations on the chart:
"bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright",
"right", or "center". Default is "bottomright". Use \code{legend.loc=NULL}
to suppress the legend.}

\item{las}{one of {0, 1, 2, 3} to set the direction of axis labels, same as
in \code{plot}. Default here is 1.}

\item{VaR.method}{a method for computing VaR; one of "modified", "gaussian",
"historical" or "kernel". VaR is computed using
\code{\link[PerformanceAnalytics]{VaR}}. Default is "historical".}

\item{loop}{logical to indicate if the plot menu should be repeated. Default
is \code{TRUE}.}

\item{...}{further arguments to be passed to other plotting functions.}
}
\description{
Generic \code{plot} method for object of class \code{tsfm}.
Plots chosen characteristic(s) for one or more assets.
}
\details{
If the plot type argument is not specified, a menu prompts for user input
and the corresponding plot is output. And, the menu is repeated for
user convenience in plotting multiple characteristics. Selecting '0' from
the menu exits the current \code{plot.tsfm} call. Alternately, setting
\code{loop=FALSE} will exit after plotting any one chosen characteristic
without the need for menu selection.

For group plots (the default), the first \code{max.show} assets are plotted.
For individual plots, \code{asset.name} is necessary if multiple assets
were modeled in \code{x} and \code{plot.single=TRUE}. However, if the
\code{fitTsfm} object contains only one asset's factor model fit,
\code{plot.tsfm} can infer this automatically, without user input.

CUSUM plots (individual asset plot options 10, 11 and 12) are applicable
only for \code{fit.method="OLS"}.

Rolling estimates (individual asset plot option 13) is not applicable for
\code{variable.slection="lars"}.
}
\examples{
# load data from the database
data(managers)
fit.macro <- fitTsfm(asset.names=colnames(managers[,(1:6)]),
                     factor.names=colnames(managers[,(7:8)]),
                     rf.name="US.3m.TR", data=managers)

# group plot(default); select type from menu prompt
# menu is auto-looped to get multiple types of plots based on the same fit
# plot(fit.macro)

# plot the factor betas of 1st 4 assets fitted above
# loop disabled to get one type of plot without interative menu
plot(fit.macro, max.show=4, which.plot.group=2, loop=FALSE)

# plot factor model return correlation; angular order of the eigenvectors
plot(fit.macro, which.plot.group=7, loop=FALSE,
     order="AOE", method="ellipse", tl.pos = "d")

# histogram of residuals from an individual asset's factor model fit
plot(fit.macro, plot.single=TRUE, asset.name="HAM1", which.plot.single=8,
     loop=FALSE)
}
\author{
Eric Zivot, Sangeetha Srinivasan and Yi-An Chen
}
\seealso{
\code{\link{fitTsfm}} and \code{\link{summary.tsfm}} for details
about the time series factor model fit, extractor functions and summary
statistics.

\code{\link[strucchange]{efp}} for CUSUM tests.

\code{\link[xts]{plot.xts}},
\code{\link[PerformanceAnalytics]{chart.TimeSeries}},
\code{\link[PerformanceAnalytics]{chart.ACFplus}},
\code{\link[PerformanceAnalytics]{chart.Histogram}},
\code{\link[PerformanceAnalytics]{chart.QQPlot}},
\code{\link[graphics]{barplot}}, \code{\link[lattice]{barchart}} and
\code{\link[corrplot]{corrplot}} for plotting methods used.

\code{\link{fmSdDecomp}}, \code{\link{fmEsDecomp}},
\code{\link{fmVaRDecomp}} for factor model risk measures.
}

