% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.tsfmUpDn.r
\name{plot.tsfmUpDn}
\alias{plot.tsfmUpDn}
\title{Plot actual against fitted values of up and down market time series factor model}
\usage{
\method{plot}{tsfmUpDn}(object, mkt.name, assets.name, ...)
}
\arguments{
\item{object}{an object of class \code{tsfmUpDn} produced by \code{fitTsfmUpDn}.}

\item{mkt.name}{The name of market returns used in \code{fitTsfmUpDn}. It is necessary
to provide the name of market returns.}

\item{assets.name}{The name of the assets used in \code{fitTsfmUpDn}. It only supports
single asset so far.}

\item{...}{Other arguments can be used in \code{plot}.}
}
\description{
Generic \code{plot} method for object of class \code{tsfmUpDn}.
}
\details{
This method plots actual values against fitted value of up and down market time series
factor model. The black dots are actual values and the red lines are fitted values.

For other types of plots, use the list objects \code{Up} and \code{Dn} of class \code{tsfmUpDn}.
The \code{plot.tsfm} can be applied.
}
\examples{
# load data from the database
data(managers)
# example: Up and down market factor model with OLS fit
fitUpDn <- fitTsfmUpDn(asset.names=colnames(managers[,(1:6)]),mkt.name="SP500.TR",
                       data=managers, fit.method="OLS",control=NULL)
 # plot the fitted model of the first asset
 plot(fitUpDn,mkt.name="SP500.TR",assets.name="HAM1")
}
\author{
Yi-An Chen
}
\seealso{
\code{\link{fitTsfmUpDn}}
}

