% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portRsq.R
\name{portRsqr}
\alias{portRsqr}
\title{R-squared and Adjusted R-squared for a Portfolio}
\usage{
portRsqr(ffmObj, weight = NULL, ...)
}
\arguments{
\item{ffmObj}{an object of class \code{ffm} produced by \code{fitFfm}}

\item{weight}{a vector of weights of the assets in the portfolio. Default is NULL.}

\item{...}{additional arguments unused}
}
\value{
\code{portRsqr} returns an object of class list with the follwing components:
\item{port.Rsqr} time series of R-squared values for the portfolio.
\item{port.AdjRsqr} time series of adjusted R-squared values for the portfolio.
}
\description{
Calcluate the R-squared and Adjusted R-squared for the portfolio of assets
}
\examples{

#Load the data 
data("stocks145scores6")
 
#Fit a Ffm
fit <- fitFfm(data=data145, asset.var="TICKER", ret.var="RETURN", 
              date.var="DATE", exposure.vars="SECTOR")
              
#Find the portfolio R-squared and adjusted portfolio R-squared for the 145 stocks data with default weights.               
portRsqr(fit)

}
\author{
Avinash Acharya
}

