% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portRsq.R
\name{portRsqr}
\alias{portRsqr}
\title{R-squared and Adjusted R-squared for a Portfolio}
\usage{
portRsqr(ffmObj, weight = NULL, rsq = T, rsqAdj = F, digits = 2, ...)
}
\arguments{
\item{ffmObj}{an object of class \code{ffm} produced by \code{fitFfm}}

\item{weight}{a vector of weights of the assets in the portfolio. Default is NULL.}

\item{rsq}{logical; if \code{TRUE}, R-squared values are computed for the portfolio. Default is \code{TRUE}.}

\item{rsqAdj}{logical; if \code{TRUE}, Adjusted R-squared values are computed for the portfolio. Default is \code{FALSE}.}

\item{digits}{an integer indicating the number of decimal places to be used for rounding. Default is 2.}

\item{...}{potentially further arguments passed.}
}
\value{
\code{portRsqr} returns a list with the follwing components:
\item{port.Rsqr}{ length-T vector of R-squared values for the portfolio.}
\item{port.AdjRsqr}{ length-T vector of adjusted R-squared values for the portfolio.}
}
\description{
Calcluate the R-squared and Adjusted R-squared for a portfolio of assets
}
\examples{

#Load the data 
data("factorDataSetDjia5Yrs")
 
#Fit a Ffm
fit <- fitFfm(data=factorDataSetDjia5Yrs, asset.var="TICKER", ret.var="RETURN", 
              date.var="DATE", exposure.vars="SECTOR")
              
#Find the portfolio R-squared values 
portRsqr(fit)



}
\author{
Avinash Acharya
}

