% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portVolDecomp.R
\name{portVolDecomp}
\alias{portVolDecomp}
\alias{portVolDecomp.tsfm}
\alias{portVolDecomp.ffm}
\title{Decompose portfolio variance risk into factor/residual risk}
\usage{
portVolDecomp(object, ...)

\method{portVolDecomp}{tsfm}(object, weights = NULL, factor.cov,
  use = "pairwise.complete.obs", ...)

\method{portVolDecomp}{ffm}(object, weights = NULL, factor.cov, ...)
}
\arguments{
\item{object}{fit object of class \code{tsfm}, or \code{ffm}.}

\item{...}{optional arguments passed to \code{\link[stats]{cov}}.}

\item{weights}{a vector of weights of the assets in the portfolio. Default is NULL, 
in which case an equal weights will be used.}

\item{factor.cov}{optional user specified factor covariance matrix with 
named columns; defaults to the sample covariance matrix.}

\item{use}{an optional character string giving a method for computing 
covariances in the presence of missing values. This must be (an 
abbreviation of) one of the strings "everything", "all.obs", 
"complete.obs", "na.or.complete", or "pairwise.complete.obs". Default is 
"pairwise.complete.obs".}
}
\value{
A vector containing: percent factor contribution to risk
portfolio volatility risk, factor volatility risk and 
residual/specific volatility risk
}
\description{
Decompose portfolio variance risk into factor/residual risk
}
\examples{
# Time Series Factor Model
data(managers)
fit.macro <- FactorAnalytics::fitTsfm(asset.names=colnames(managers[,(1:6)]),
                     factor.names=colnames(managers[,(7:9)]),
                     rf.name=colnames(managers[,10]), data=managers)
decomp <- portVolDecomp(fit.macro)
decomp

# Fundamental Factor Model
data("stocks145scores6")
dat = stocks145scores6
dat$DATE = as.yearmon(dat$DATE)
dat = dat[dat$DATE >=as.yearmon("2008-01-01") & 
          dat$DATE <= as.yearmon("2012-12-31"),]

# Load long-only GMV weights for the return data
data("wtsStocks145GmvLo")
wtsStocks145GmvLo = round(wtsStocks145GmvLo,5)  
                                                     
# fit a fundamental factor model
fit.cross <- fitFfm(data = dat, 
              exposure.vars = c("SECTOR","ROE","BP","MOM121","SIZE","VOL121",
              "EP"),date.var = "DATE", ret.var = "RETURN", asset.var = "TICKER", 
              fit.method="WLS", z.score = "crossSection")
              
decomp = portVolDecomp(fit.cross) 
# get the factor contributions of risk 
decomp             
 
}
\seealso{
\code{\link{fitTsfm}}, \code{\link{fitFfm}}
for the different factor model fitting functions.

\code{\link{portSdDecomp}} for portfolio factor model VaR decomposition.
\code{\link{portVaRDecomp}} for portfolio factor model VaR decomposition.
\code{\link{portEsDecomp}} for portfolio factor model ES decomposition.
}
\author{
Douglas Martin, Lingjie Yi
}
