% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.pafm.r
\name{print.pafm}
\alias{print.pafm}
\title{Print object of class \code{"pafm"}.}
\usage{
\method{print}{pafm}(x, ...)
}
\arguments{
\item{x}{object of class \code{"pafm"} created by
\code{paFm}.}

\item{...}{Other arguments for \code{print} methods.}
}
\description{
Generic function of print method for \code{paFm}.
}
\examples{
# load data from the database
 data(managers)
# fit the factor model with LS
fit <- fitTsfm(asset.names=colnames(managers[,(1:6)]),
               factor.names=c("EDHEC.LS.EQ","SP500.TR"), data=managers)
fm.attr <- paFm(fit)
print(fm.attr)
}
\author{
Yi-An Chen.
}

