% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.sfm.r
\name{print.sfm}
\alias{print.sfm}
\title{Prints out a fitted statictical factor model object}
\usage{
\method{print}{sfm}(x, digits = max(3, .Options$digits - 3), ...)
}
\arguments{
\item{x}{an object of class \code{sfm} produced by \code{fitSfm}.}

\item{digits}{an integer value, to indicate the required number of
significant digits. Default is 3.}

\item{...}{optional arguments passed to the \code{print} method.}
}
\description{
S3 \code{print} method for object of class \code{sfm}. Prints
the call, factor model dimension, factor loadings, r-squared and residual
volatilities from the fitted object.
}
\examples{
data(StockReturns)
fit <- fitSfm(r.M, k=2)
print(fit)
}
\author{
Yi-An Chen and Sangeetha Srinivasan
}
\seealso{
\code{\link{fitSfm}}, \code{\link{summary.sfm}}
}

