% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadDataGitHub.R
\name{readGithubFAdata}
\alias{readGithubFAdata}
\title{download FA dataset from github}
\usage{
readGithubFAdata(filename = "")
}
\arguments{
\item{filename}{stockCRSPdaily, stockCRSPweekly, stockCRSPmonthly}
}
\value{

}
\description{
download FA dataset from github
}
\examples{
readGithubFAdata(filename  = "stockCRSPweekly")

}
\author{
Kirk Li.
}
