% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repReturn.R
\name{repReturn}
\alias{repReturn}
\title{Portfolio return decomposition report}
\usage{
repReturn(ffmObj, weights = NULL, isPlot = TRUE, isPrint = TRUE,
  stripLeft = TRUE, ...)
}
\arguments{
\item{ffmObj}{an object of class ffm returned by fitFfm.}

\item{isPlot}{logical variable to generate plot or not.}

\item{isPrint}{logical variable to print numeric summary or not.}

\item{stripLeft}{logical variable to choose the position of strip, "TRUE" for drawing strips on the left of each panel, "FALSE" for drawing strips on the top of each panel}

\item{...}{futher arguments passed to or from other methods.}

\item{weight}{a vector of weights of the assets in the portfolio. Default is NULL.}
}
\description{
Decompostite return of portfolio into return of different factors based on fundamental factor model. This method takes fundamental factor model fit, "ffm" object, and portfolio weight as inputs and generates numeric summary and plot visualization.
}
\examples{

#Load fundamental and return data 
data("stocks145scores6")
dat = stocks145scores6
dat$DATE = as.yearmon(dat$DATE)
dat = dat[dat$DATE >=as.yearmon("2008-01-01") & dat$DATE <= as.yearmon("2012-12-31"),]

#Load long-only GMV weights for the return data
data("wtsStocks145GmvLo")
                                                     
#fit a fundamental factor model
fit <- fitFfm(data = dat, 
              exposure.vars = c("SECTOR","ROE","BP","PM12M1M","SIZE","ANNVOL1M","EP"),
              date.var = "DATE", ret.var = "RETURN", asset.var = "TICKER", 
              fit.method="WLS", z.score = T)

repReturn(fit, wtsStocks145GmvLo)               
}
\author{
Lingjie Yi
}

