% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsqTstatsTs.R
\name{rsqTstatsTs}
\alias{rsqTstatsTs}
\title{Plots from a fitted fundamental factor model}
\usage{
rsqTstatsTs(ffmObj, col = "cyan", z.score = 1.96, ...)
}
\arguments{
\item{ffmObj}{an object of class \code{ffm} produced by \code{fitFfm}}

\item{col}{A specification for the default plotting color. Default is cyan}

\item{z.score}{critical value corresponding to the confidence interval. Default is 1.96 i.e 95\% C.I}

\item{...}{potentially further arguments passed.}
}
\description{
Plot time series of the R-squared values, t-statistic vales and the
 number of risk indices with significant t-stats for a fundamentally fit object. By default, all the three plots are plotted.
}
\examples{

#Load the data 
data("factorDataSetDjia5Yrs")
 
#Fit a Ffm
fit <- fitFfm(data=factorDataSetDjia5Yrs, asset.var="TICKER", ret.var="RETURN", 
              date.var="DATE", exposure.vars="SECTOR")
              riskindices.mod <- fitFfm(data = stacked.df,exposure.vars = c("MARKETCAP","ENTVALUE","P2B","EV2S"),
              date.var = "DATE", ret.var = "RETURN", asset.var = "TICKER", fit.method="WLS")
   
 rsqTstatsTs(fit)  
   
              

}
\author{
Doug Martin, Avinash Acharya
}

