% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFfmDT.R
\name{standardizeExposures}
\alias{standardizeExposures}
\title{standardizeExposures}
\usage{
standardizeExposures(
  specObj,
  Std.Type = c("None", "CrossSection", "TimeSeries"),
  lambda = 0.9
)
}
\arguments{
\item{specObj}{is a ffmSpec object,}

\item{Std.Type}{method for exposure standardization; one of "none",
"CrossSection", or "TimeSeries".
Default is \code{"none"}.}

\item{lambda}{lambda value to be used for the EWMA estimation of residual 
variances. Default is 0.9}
}
\value{
the ffM spec object with exposures z-scored
}
\description{
function to calculate z-scores for numeric exposure using weights weight.var
}
\details{
this function operates on the data inside the specObj and applies a 
standardization to it.  The user can choose CrossSectional or timeSeries standardization
}
\seealso{
\code{\link{specFfm()}} for information on the definition of the specFfm object.
}
