% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFfmDT.R
\name{standardizeReturns}
\alias{standardizeReturns}
\title{standardizeReturns}
\usage{
standardizeReturns(
  specObj,
  GARCH.params = list(omega = 0.09, alpha = 0.1, beta = 0.81)
)
}
\arguments{
\item{specObj}{is a ffmSpec object}

\item{GARCH.params}{fixed Garch(1,1) parameters}
}
\value{
an ffmSpec Object with the standardized returns added
}
\description{
Standardize the returns using GARCH(1,1) volatilities.
}
\details{
this function operates on the data inside the specObj and standardizes
the returns to create scaled return.
}
\seealso{
\code{\link{specFfm}} for information on the definition of the specFfm object.
}
