\name{summary.FM.attribution}
\alias{summary.FM.attribution}
\title{summary FM.attribution object.}
\usage{
  summary.FM.attribution(fm.attr)
}
\arguments{
  \item{fm.attr}{FM.attribution object created by
  factorModelPerformanceAttribution.}
}
\description{
  Generic function of summary method for
  factorModelPerformanceAttribution.
}
\examples{
\dontrun{
  fm.attr <- factorModelPerformanceAttribution(fit.macro)
  summary(fm.attr)
  }
}
\author{
  Yi-An Chen.
}

