% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.pafm}
\alias{summary.pafm}
\title{summary \code{"pafm"} object.}
\usage{
\method{summary}{pafm}(object, digits = max(3, .Options$digits - 3), ...)
}
\arguments{
\item{object}{\code{"pafm"} object created by
\code{paFm}.}

\item{digits}{integer indicating the number of decimal places. Default is 3.}

\item{...}{Other arguments for \code{print} methods.}
}
\description{
Generic function of summary method for \code{paFm}.
}
\examples{
# load data from the database
data(managers)
# fit the factor model with OLS
fit.ts <- fitTsfm(asset.names=colnames(managers[,(1:6)]),
                  factor.names=c("EDHEC LS EQ","SP500 TR"),
                  mkt.name="SP500 TR", mkt.timing="both", data=managers)

fm.attr <- paFm(fit.ts)
summary(fm.attr)
}
\author{
Yi-An Chen.
}

