% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.tsfm}
\alias{print.summary.tsfm}
\alias{summary.tsfm}
\title{Summarizing a fitted time series factor model}
\usage{
\method{summary}{tsfm}(object, ...)

\method{print}{summary.tsfm}(x, digits = 3, ...)
}
\arguments{
\item{object}{an object of class \code{tsfm} returned by \code{fitTSFM}.}

\item{x}{an object of class \code{summary.tsfm}.}

\item{digits}{number of significants digits to use when printing.
Default is 3.}

\item{...}{futher arguments passed to or from other methods.}
}
\value{
Returns an object of class \code{summary.tsfm}, which is a list
containing the function call to \code{fitTSFM} and the
\code{summary.lm} objects fitted for each asset in the factor model.
The print method for class \code{summary.tsfm} outputs the call,
coefficients, r-squared and residual volatilty for all assets.
}
\description{
\code{summary} method for object of class \code{tsfm}.
Returned object is of class {summary.tsfm}.
}
\note{
For a more detailed printed summary for each asset, refer to
\code{print.summary.lm}, which further formats the coefficients,
standard errors, etc. and additionally gives significance
stars if \code{signif.stars} is TRUE.
}
\examples{
data(managers.df)
fit <- fitTSFM(asset.names=colnames(managers.df[,(1:6)]),
               factor.names=colnames(managers.df[,7:9]),
               market.name="SP500.TR", data=managers.df,
               fit.method="OLS", variable.selection="none",
               add.up.market=TRUE, add.market.sqd=TRUE)
# summary of factor model fit for all assets
summary(fit)

# summary of lm fit for a single asset
summary(fit$asset.fit[[1]])
}
\author{
Yi-An Chen & Sangeetha Srinivasan.
}
\seealso{
\code{\link{fitTSFM}}, \code{\link[stats]{summary.lm}}
}

