% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.tsfm}
\alias{summary.tsfm}
\title{Summarizing a fitted time series factor model}
\usage{
summary.tsfm(object, digits = 3, ...)
}
\arguments{
\item{object}{an object of class \code{tsfm} produced by \code{fitTSFM}.}

\item{digits}{an integer value, to indicate the required number of
significant digits. Default is 3.}

\item{...}{optional arguments passed to the \code{print} method.}
}
\value{
Returns an object of class {summary.tsfm}.
}
\description{
S3 \code{summary} method for object of class \code{tsfm}.
Resulting object is of class {summary.tsfm}.
}
\examples{
data(managers.df)
fit <- fitTSFM(asset.names=colnames(managers.df[,(1:6)]),
               factor.names=colnames(managers.df[,7:9]),
               market.name="SP500.TR",
               data=data, fit.method="OLS", variable.selection="none",
               add.up.market=TRUE, add.market.sqd=TRUE)
summary(fit)
}
\author{
Yi-An Chen & Sangeetha Srinivasan.
}
\seealso{
\code{\link{fitTSFM}}, \code{\link{tsfm}}
}

