% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.tsfm}
\alias{print.summary.tsfm}
\alias{summary.tsfm}
\title{Summarizing a fitted time series factor model}
\usage{
\method{summary}{tsfm}(object, se.type = "Default", ...)

\method{print}{summary.tsfm}(x, digits = 3, ...)
}
\arguments{
\item{object}{an object of class \code{tsfm} returned by \code{fitTsfm}.}

\item{se.type}{one of "Default", "HC" or "HAC"; option for computing
HC/HAC standard errors and t-statistics.}

\item{x}{an object of class \code{summary.tsfm}.}

\item{digits}{number of significants digits to use when printing.
Default is 3.}

\item{...}{futher arguments passed to or from other methods.}
}
\value{
Returns an object of class \code{summary.tsfm}.
The print method for class \code{summary.tsfm} outputs the call,
coefficients (with standard errors and t-statistics), r-squared and
residual volatilty (under the homoskedasticity assumption) for all assets.

Object of class \code{summary.tsfm} is a list of length N + 2 containing:
\item{call}{the function call to \code{fitTsfm}}
\item{se.type}{standard error type as input}
\item{}{summaries of the N fit objects (of class \code{lm}, \code{lmRob}
or \code{lars}) for each asset in the factor model.}
}
\description{
\code{summary} method for object of class \code{tsfm}.
Returned object is of class {summary.tsfm}.
}
\details{
The default \code{summary} method for a fitted \code{lm} object
computes the standard errors and t-statistics under the assumption of
homoskedasticty. Argument \code{se.type} gives the option to compute
heteroskedasticity-consistent (HC) or
heteroskedasticity-autocorrelation-consistent (HAC) standard errors and
t-statistics using \code{\link[lmtest]{coeftest}}. This option is meaningful
only if \code{fit.method = "OLS" or "DLS"}. This option is currently not
available for \code{variable.selection = "lar" or "lasso"}.
}
\note{
For a more detailed printed summary for each asset, refer to
\code{\link[stats]{summary.lm}} or \code{\link[robust]{lmRob}}, which
include F-statistics, Multiple R-squared, Adjusted R-squared and further
format the coefficients, standard errors, etc. and additionally give
significance stars if \code{signif.stars} is TRUE.
}
\examples{
data(managers)
fit <- fitTsfm(asset.names=colnames(managers[,(1:6)]),
               factor.names=colnames(managers[,7:9]),
               mkt.name="SP500 TR", mkt.timing="HM", data=managers)

# summary of factor model fit for all assets
summary(fit, "HAC")

# summary of lm fit for a single asset
summary(fit$asset.fit[[1]])
}
\author{
Yi-An Chen & Sangeetha Srinivasan.
}
\seealso{
\code{\link{fitTsfm}}, \code{\link[stats]{summary.lm}}
}

