% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsPlotMP.R
\name{tsPlotMP}
\alias{tsPlotMP}
\title{Time Series Plots}
\usage{
tsPlotMP(
  ret,
  add.grid = FALSE,
  layout = NULL,
  type = "l",
  yname = "RETURNS (\%)",
  Pct = F,
  scaleType = "free",
  stripLeft = TRUE,
  main = NULL,
  lwd = 1,
  stripText.cex = 1,
  axis.cex = 1,
  color = "black",
  zeroLine = TRUE
)
}
\arguments{
\item{add.grid}{logical varible.If 'TRUE', type = c('l', 'g'); If 'FALSE', type = c('l')}

\item{layout}{layout is a numeric vector of length 2 or 3 giving the number of columns, rows, and pages (optional) in a multipanel display.}

\item{type}{character. type of the plot; 'l' denotes a line, 'p' denotes a point, and 'b' and 'o' both denote both together.deafault is 'l'.}

\item{yname}{character or espression giving label(s) for the y-axis}

\item{Pct}{Pct controls if use the percentage value.}

\item{scaleType}{scaleType controls if use a same scale of y-axis, choose from c('same', 'free')}

\item{stripLeft}{logical variable to choose the position of strip, 'TRUE' for drawing strips on the left of each panel, 'FALSE' for drawing strips on the top of each panel}

\item{main}{Typically a character string or expression describing the main title.}

\item{lwd}{The line width, a positive number, defaulting to 1}

\item{stripText.cex}{a number indicating the amount by which strip text in the plot(s) should be scaled relative to the default. 1=default, 1.5 is 50\% larger, 0.5 is 50\% smaller, etc.}

\item{axis.cex}{a number indicating the amount by which axis in the plot(s) should be scaled relative to the default. 1=default, 1.5 is 50\% larger, 0.5 is 50\% smaller, etc.}

\item{color}{A specification for the default plotting color. Default is black.}

\item{zeroLine}{logical varible to choose add a dotted horizontal line at the zero vertical distance}

\item{data}{an time series exposure/return object}
}
\description{
Plot time series with specific plotting parameters
}
\examples{

#Load the data
data("stocks145scores6")
dat = stocks145scores6
returns = tapply(dat$RETURN,list(dat$DATE,dat$TICKER),I)
ret = xts::xts(returns[,1:5],as.yearmon(rownames(returns)))

#generate return time series plot               
tsPlotMP(ret, color = 'Blue')
tsPlotMP(ret, scaleType = "same", zeroLine = FALSE)
tsPlotMP(ret, stripLeft = FALSE, main = 'Time Series Plot')
tsPlotMP(ret, stripLeft = FALSE, main = 'Time Series Plot', layout = c(3,3))
 
   
}
\author{
Douglas Martin, Lingjie Yi
}
