% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsPlotMP.R
\name{tsPlotMP}
\alias{tsPlotMP}
\title{Plot}
\usage{
tsPlotMP(ret, add.grid = F, cex = 1, layout = NULL, type = "l",
  pct = 100, yname = "RETURNS (\%)", scaleType = "free", stripLeft = T,
  main = NULL, lwd = 1, color = "black")
}
\description{

}

