% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wtsDjiaGmv}
\alias{wtsDjiaGmv}
\title{DJIA GMV portfolio weights}
\source{
TBA
}
\usage{
data("wtsDjiaGmv")
}
\description{
Contains weights obtained after optimizing the portfolio returns of the 30 DJIA stocks (from dataset factorDataSetDjia5Yrs)
             for a global minimum variance portfolio starting from Jan 2008 to Dec 2012.
}
