% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wtsDjiaGmvLo}
\alias{wtsDjiaGmvLo}
\title{DJIA GMV long-only portfolio weights}
\source{
TBA
}
\usage{
data("wtsDjiaGmvLo")
}
\description{
Contains weights obtained after optimizing the portfolio returns of the 30 DJIA stocks (from dataset factorDataSetDjia5Yrs) 
             for a long-only global minimum variance portfolio starting from Jan 2008 to Dec 2012.
}
