% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wtsStocks145GmvLo}
\alias{wtsStocks145GmvLo}
\title{CRSP 145 stocks GMV long-only weights}
\source{
TBA
}
\usage{
data("wtsStocks145GmvLo")
}
\description{
Contains weights obtained after optimizing the portfolio returns of 145 stocks (from dataset stocks145scores6)
             for a long-only global minimum variance portfolio starting from Jan 1990 to Dec 2014.
}
