% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtsStocks145GmvLo.R
\docType{data}
\name{wtsStocks145GmvLo}
\alias{wtsStocks145GmvLo}
\title{wtsStocks145GmvLo}
\format{
A vector of type numeric with 145 observations
}
\source{
TBA
}
\usage{
data('wtsStocks145GmvLo')
}
\description{
Contains weights obtained after optimizing the portfolio returns of 145 stocks (from dataset \link{stocks145scores6})
             for a long-only global minimum variance portfolio starting from Jan 1990 to Dec 2014.
}
\details{
CRSP 145 stocks GMV long-only weights
}
\examples{
 
data(wtsStocks145GmvLo)
 str(wtsStocks145GmvLo)
}
\keyword{datasets}
